Repository

Repository stores specific values of an artifact, different for various artifact's classes and artifacts themselves. To learn more about artifacts visit archivist-package.

Details

Repository is a folder with an SQLite database stored in a file named backpack and a subdirectory named gallery.

backpack contains two tables:artifact and tag. artifact table consists of three columns:

md5hash,

name,

createdDate,

while tag table consists of the following three columns:

artifact,

tag,

createdDate.

gallery collects the following objects:

artifacts and artifacts' data saved as .rda files,

artifacts' miniatures saved as .txt and .png files.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results URL https://github.com/pbiecek/archivist

See also

Functions using Repository are:

addTagsRepo,

ahistory,

aread,

asearch,

cache,

getTagsLocal,

getTagsRemote,

splitTagsLocal,

splitTagsRemote,

loadFromLocalRepo,

loadFromRemoteRepo,

rmFromLocalRepo,

saveToRepo,

searchInLocalRepo,

searchInRemoteRepo,

shinySearchInLocalRepo,

showLocalRepo,

showRemoteRepo,

summaryLocalRepo,

summaryRemoteRepo.

Function creating Repository is:

createLocalRepo.

Function deleting Repository is:

deleteLocalRepo.

Functions coping Repository are:

copyLocalRepo,

copyRemoteRepo.

Functions creating a zip archive from an existing Repository are:

zipLocalRepo,

zipRemoteRepo.

Functions setting global path to the Repository are:

setLocalRepo,

setRemoteRepo.

Learn more about Repository at archivist wiki webpage on https://github.com/pbiecek/archivist/wiki/archivist-package-Repository.

Other archivist: Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo