Show Artifact's List of Foramts

aformat extracts artifact's formats. Having formats one may decide which should he read. Currently only rda format is supported for artifact and txt/png for miniatures.

aformat(md5hash = NULL)

Arguments

md5hash

One of the following (see aread): A character vector which elements are consisting of at least three components separated with '/': Remote user name, Remote repository and name of the artifact (MD5 hash) or it's abbreviation. MD5 hashes of artifacts in current local default directory or its abbreviations.

Value

A vector of characters.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

## Not run: ------------------------------------
# setLocalRepo(system.file("graphGallery", package = "archivist"))
# aformat("2a6e492cb6982f230e48cf46023e2e4f")
# 
# # old
# aformat("pbiecek/graphGallery/2a6e492cb6982f230e48cf46023e2e4f")
# # png
# aformat("pbiecek/graphGallery/7f3453331910e3f321ef97d87adb5bad")
## ---------------------------------------------