Tools for Storing, Restoring and Searching for R Objects

Data exploration and modelling is a process in which a lot of data artifacts are produced. Artifacts like: subsets, data aggregates, plots, statistical models, different versions of data sets and different versions of results. The more projects we work on, the more artifacts are produced and the harder it is to manage these artifacts.

Archivist helps to store and manage artifacts created in R.

Archivist allows you to store selected artifacts as binary files along with their metadata and relations. Archivist allows you to share artifacts with others, either through a shared folder or github. Archivist allows you to look for artifacts by using its class, name, date of creation or other properties. It also facilitates restoring such artifacts. Archivist allows you to check if a new artifact is the exact copy of the one that was produced some time ago. This might be useful either for testing or caching.

The list of main use cases is available here https://github.com/pbiecek/archivist.

Extensions of archivist are

Tools for Archiving, Managing and Sharing R Objects via GitHub: archive marcinkosinski.github.io/archivist.github/

Details

For more detailed information visit archivist wiki on Github.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

Blogging

We have prepared history of blog posts and conference talks about archivist under this link http://pbiecek.github.io/archivist/articles/posts.html

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results Conditionally accepted for Journal of Statistical Software URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, areadLocal, aread, asearchLocal, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo