Read Artifacts Given as md5hashes from the Repository

areadLocal reads the artifact from the Local Repository. It's a wrapper around loadFromLocalRepo.

areadLocal(md5hash, repo)

Arguments

md5hash
A character vector which elements are consisting name of the repository and name of the artifact (MD5 hash) or it's abbreviation.

repo
A character with path to local repository.

Details

Function areadLocal reads artifacts (by md5hashes) from Local Repository.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Biecek P and Kosinski M (2017). "archivist: An R Package for Managing, Recording and Restoring Data Analysis Results." _Journal of Statistical Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL: http://doi.org/10.18637/jss.v082.i11). URL https://github.com/pbiecek/archivist

See also