Read Artifacts Given as a List of Tags

asearchLocal searches for artifacts that contain all specified Tags and reads all of them from a local Repository. It's a wrapper around searchInLocalRepo and loadFromLocalRepo.

asearchLocal(patterns, repo = NULL)

Arguments

patterns
A character vector of Tags. Only artifacts that contain all Tags are returned.

repo
A character with path to local repository.

Value

This function returns a list of artifacts (by their values).

Details

Function asearchLocal reads all artifacts that contain given list of Tags from the selected Local Repository.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Biecek P and Kosinski M (2017). "archivist: An R Package for Managing, Recording and Restoring Data Analysis Results." _Journal of Statistical Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL: http://doi.org/10.18637/jss.v082.i11). URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, areadLocal, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo