Add Tracing For All Objects Created By Given Function

atrace add call to saveToLocalRepo at the end of a given function.

atrace(FUN = "lm", object = "z")

Arguments

FUN

name of a function to be traced (character)

object

name of an object that should be traced (character)

Details

Function atrace calls the trace function.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

# read the object from local directory
createLocalRepo("arepo_test", default=TRUE)
atrace("lm", "z")

#> Tracing function "lm" in package "stats"

#> [1] "lm"

lm(Sepal.Length~Sepal.Width, data=iris)

#> Tracing lm(Sepal.Length ~ Sepal.Width, data = iris) on exit 

#> 
#> Call:
#> lm(formula = Sepal.Length ~ Sepal.Width, data = iris)
#> 
#> Coefficients:
#> (Intercept)  Sepal.Width  
#>      6.5262      -0.2234  
#> 

asearch("class:lm")

#> $`3944d17081ad278b3b435857826b199a`
#> 
#> Call:
#> lm(formula = Sepal.Length ~ Sepal.Width, data = iris)
#> 
#> Coefficients:
#> (Intercept)  Sepal.Width  
#>      6.5262      -0.2234  
#> 
#> 

untrace("lm")

#> Untracing function "lm" in package "stats"