Return Tags Corresponding to md5hash

getTagsLocal and getTagsRemote return Tags (see Tags) related to md5hash of an artifact. To learn more about artifacts visit archivist-package.

getTagsLocal(md5hash, repoDir = aoptions("repoDir"), tag = "name")

getTagsRemote(md5hash, repo = aoptions("repo"), user = aoptions("user"),
  branch = aoptions("branch"), subdir = aoptions("subdir"),
  repoType = aoptions("repoType"), tag = "name")

Arguments

md5hash
A character containing md5hash of artifacts which Tags are desired to be returned.

repoDir
A character denoting an existing directory in which artifacts are stored.

tag
A regular expression denoting type of a Tag that we search for (see Examples). Default tag = "name".

repo
While working with the Remote repository. A character containing a name of the Remote repository on which the Repository is stored. By default set to NULL - see Note.

user
While working with the Remote repository. A character containing a name of the Remote user on whose account the repo is created. By default set to NULL - see Note.

branch
While working with the Remote repository. A character containing a name of the Remote repository's branch on which the Repository is stored. Default branch is master.

subdir
While working with the Remote repository. A character containing a name of a directory on the Remote repository on which the Repository is stored. If the Repository is stored in main folder on the Remote repository, this should be set to subdir = "/" as default.

repoType
A character containing a type of the remote repository. Currently it can be 'github' or 'bitbucket'.

Value

The character vector of Tags (see Tags) related to md5hash of an artifact.

Details

getTagsLocal and getTagsRemote return Tags, of a specific type described by tag parameter, related to md5hash of an artifact. To learn more about artifacts visit archivist-package.

Note

If repo and user are set to NULL (as default) in Remote mode then global parameters set in setRemoteRepo function are used.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Biecek P and Kosinski M (2017). "archivist: An R Package for Managing, Recording and Restoring Data Analysis Results." _Journal of Statistical Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL: http://doi.org/10.18637/jss.v082.i11). URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, areadLocal, aread, asearchLocal, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, areadLocal, aread, asearchLocal, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

### Local version

## EXAMPLE with pipe operator %a%

# Creating empty repository
exampleRepoDir <- tempfile()
createLocalRepo( exampleRepoDir )

library(dplyr)
data(mtcars)
setLocalRepo(repoDir = exampleRepoDir)
hash <- mtcars %a%
 group_by(cyl, am) %a%
   select(mpg, cyl, wt, am) %a%
     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a%
       filter(avgmpg > 20) %a%
       saveToRepo( exampleRepoDir )

showLocalRepo(exampleRepoDir)

#>                             md5hash
#> 1  a63c70e73b58d0823ab3bcbd3b543d6f
#> 2  7e3aefa426aa39c40a63405b6dfec950
#> 3  403c3921f0fb68cbc8d5c6ca18e49141
#> 4  7e3aefa426aa39c40a63405b6dfec950
#> 5  403c3921f0fb68cbc8d5c6ca18e49141
#> 6  7e3aefa426aa39c40a63405b6dfec950
#> 7  05c7067c4f4fe4fcb5fad606097815c1
#> 8  7e3aefa426aa39c40a63405b6dfec950
#> 9  05c7067c4f4fe4fcb5fad606097815c1
#> 10 7e3aefa426aa39c40a63405b6dfec950
#> 11 497ae8d6e1657de7bfbb995b3d0555d0
#> 12 7e3aefa426aa39c40a63405b6dfec950
#> 13 d6c834d34a9be232e2ada1455143cac6
#> 14 7e3aefa426aa39c40a63405b6dfec950
#> 15 d87c917f988f8035157eeac8c1acb1a3
#> 16 7e3aefa426aa39c40a63405b6dfec950
#> 17 d87c917f988f8035157eeac8c1acb1a3
#> 18 7e3aefa426aa39c40a63405b6dfec950
#> 19 d87c917f988f8035157eeac8c1acb1a3
#> 20 7e3aefa426aa39c40a63405b6dfec950
#> 21 61151f9db7da9f6e15f485cf083d4f89
#> 22 7e3aefa426aa39c40a63405b6dfec950
#>                                                                                                                                             name
#> 1                                                                                                                                         mtcars
#> 2                                                                                                               7e3aefa426aa39c40a63405b6dfec950
#> 3                                                                                                                                        res_val
#> 4                                                                                                               7e3aefa426aa39c40a63405b6dfec950
#> 5                                                                                                                   mtcars %a% group_by(cyl, am)
#> 6                                                                                                               7e3aefa426aa39c40a63405b6dfec950
#> 7                                                                                                                                        res_val
#> 8                                                                                                               7e3aefa426aa39c40a63405b6dfec950
#> 9                                                                                      mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am)
#> 10                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#> 11                                                                                                                                       res_val
#> 12                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#> 13                             mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt))
#> 14                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#> 15                                                                                                                                       res_val
#> 16                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#> 17                                                                                                                                         __LHS
#> 18                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#> 19 mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a% filter(avgmpg >     20)
#> 20                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#> 21                                                                                                                                       res_val
#> 22                                                                                                              7e3aefa426aa39c40a63405b6dfec950
#>            createdDate
#> 1  2017-11-25 01:47:16
#> 2  2017-11-25 01:47:16
#> 3  2017-11-25 01:47:16
#> 4  2017-11-25 01:47:16
#> 5  2017-11-25 01:47:16
#> 6  2017-11-25 01:47:16
#> 7  2017-11-25 01:47:16
#> 8  2017-11-25 01:47:17
#> 9  2017-11-25 01:47:17
#> 10 2017-11-25 01:47:17
#> 11 2017-11-25 01:47:17
#> 12 2017-11-25 01:47:17
#> 13 2017-11-25 01:47:17
#> 14 2017-11-25 01:47:17
#> 15 2017-11-25 01:47:17
#> 16 2017-11-25 01:47:17
#> 17 2017-11-25 01:47:17
#> 18 2017-11-25 01:47:17
#> 19 2017-11-25 01:47:17
#> 20 2017-11-25 01:47:17
#> 21 2017-11-25 01:47:17
#> 22 2017-11-25 01:47:17

showLocalRepo(exampleRepoDir, method = "tags")

#>                             artifact
#> 1   a63c70e73b58d0823ab3bcbd3b543d6f
#> 2   a63c70e73b58d0823ab3bcbd3b543d6f
#> 3   a63c70e73b58d0823ab3bcbd3b543d6f
#> 4   a63c70e73b58d0823ab3bcbd3b543d6f
#> 5   a63c70e73b58d0823ab3bcbd3b543d6f
#> 6   a63c70e73b58d0823ab3bcbd3b543d6f
#> 7   a63c70e73b58d0823ab3bcbd3b543d6f
#> 8   a63c70e73b58d0823ab3bcbd3b543d6f
#> 9   a63c70e73b58d0823ab3bcbd3b543d6f
#> 10  a63c70e73b58d0823ab3bcbd3b543d6f
#> 11  a63c70e73b58d0823ab3bcbd3b543d6f
#> 12  a63c70e73b58d0823ab3bcbd3b543d6f
#> 13  a63c70e73b58d0823ab3bcbd3b543d6f
#> 14  a63c70e73b58d0823ab3bcbd3b543d6f
#> 15  a63c70e73b58d0823ab3bcbd3b543d6f
#> 16  7e3aefa426aa39c40a63405b6dfec950
#> 17  a63c70e73b58d0823ab3bcbd3b543d6f
#> 18  a63c70e73b58d0823ab3bcbd3b543d6f
#> 19  403c3921f0fb68cbc8d5c6ca18e49141
#> 20  403c3921f0fb68cbc8d5c6ca18e49141
#> 21  403c3921f0fb68cbc8d5c6ca18e49141
#> 22  403c3921f0fb68cbc8d5c6ca18e49141
#> 23  403c3921f0fb68cbc8d5c6ca18e49141
#> 24  403c3921f0fb68cbc8d5c6ca18e49141
#> 25  403c3921f0fb68cbc8d5c6ca18e49141
#> 26  403c3921f0fb68cbc8d5c6ca18e49141
#> 27  403c3921f0fb68cbc8d5c6ca18e49141
#> 28  403c3921f0fb68cbc8d5c6ca18e49141
#> 29  403c3921f0fb68cbc8d5c6ca18e49141
#> 30  403c3921f0fb68cbc8d5c6ca18e49141
#> 31  403c3921f0fb68cbc8d5c6ca18e49141
#> 32  403c3921f0fb68cbc8d5c6ca18e49141
#> 33  403c3921f0fb68cbc8d5c6ca18e49141
#> 34  403c3921f0fb68cbc8d5c6ca18e49141
#> 35  403c3921f0fb68cbc8d5c6ca18e49141
#> 36  403c3921f0fb68cbc8d5c6ca18e49141
#> 37  403c3921f0fb68cbc8d5c6ca18e49141
#> 38  403c3921f0fb68cbc8d5c6ca18e49141
#> 39  7e3aefa426aa39c40a63405b6dfec950
#> 40  403c3921f0fb68cbc8d5c6ca18e49141
#> 41  403c3921f0fb68cbc8d5c6ca18e49141
#> 42  403c3921f0fb68cbc8d5c6ca18e49141
#> 43  403c3921f0fb68cbc8d5c6ca18e49141
#> 44  403c3921f0fb68cbc8d5c6ca18e49141
#> 45  403c3921f0fb68cbc8d5c6ca18e49141
#> 46  403c3921f0fb68cbc8d5c6ca18e49141
#> 47  403c3921f0fb68cbc8d5c6ca18e49141
#> 48  403c3921f0fb68cbc8d5c6ca18e49141
#> 49  403c3921f0fb68cbc8d5c6ca18e49141
#> 50  403c3921f0fb68cbc8d5c6ca18e49141
#> 51  403c3921f0fb68cbc8d5c6ca18e49141
#> 52  403c3921f0fb68cbc8d5c6ca18e49141
#> 53  403c3921f0fb68cbc8d5c6ca18e49141
#> 54  403c3921f0fb68cbc8d5c6ca18e49141
#> 55  403c3921f0fb68cbc8d5c6ca18e49141
#> 56  403c3921f0fb68cbc8d5c6ca18e49141
#> 57  403c3921f0fb68cbc8d5c6ca18e49141
#> 58  403c3921f0fb68cbc8d5c6ca18e49141
#> 59  403c3921f0fb68cbc8d5c6ca18e49141
#> 60  7e3aefa426aa39c40a63405b6dfec950
#> 61  403c3921f0fb68cbc8d5c6ca18e49141
#> 62  403c3921f0fb68cbc8d5c6ca18e49141
#> 63  05c7067c4f4fe4fcb5fad606097815c1
#> 64  05c7067c4f4fe4fcb5fad606097815c1
#> 65  05c7067c4f4fe4fcb5fad606097815c1
#> 66  05c7067c4f4fe4fcb5fad606097815c1
#> 67  05c7067c4f4fe4fcb5fad606097815c1
#> 68  05c7067c4f4fe4fcb5fad606097815c1
#> 69  05c7067c4f4fe4fcb5fad606097815c1
#> 70  05c7067c4f4fe4fcb5fad606097815c1
#> 71  05c7067c4f4fe4fcb5fad606097815c1
#> 72  05c7067c4f4fe4fcb5fad606097815c1
#> 73  05c7067c4f4fe4fcb5fad606097815c1
#> 74  05c7067c4f4fe4fcb5fad606097815c1
#> 75  05c7067c4f4fe4fcb5fad606097815c1
#> 76  7e3aefa426aa39c40a63405b6dfec950
#> 77  05c7067c4f4fe4fcb5fad606097815c1
#> 78  05c7067c4f4fe4fcb5fad606097815c1
#> 79  05c7067c4f4fe4fcb5fad606097815c1
#> 80  05c7067c4f4fe4fcb5fad606097815c1
#> 81  05c7067c4f4fe4fcb5fad606097815c1
#> 82  05c7067c4f4fe4fcb5fad606097815c1
#> 83  05c7067c4f4fe4fcb5fad606097815c1
#> 84  05c7067c4f4fe4fcb5fad606097815c1
#> 85  05c7067c4f4fe4fcb5fad606097815c1
#> 86  05c7067c4f4fe4fcb5fad606097815c1
#> 87  05c7067c4f4fe4fcb5fad606097815c1
#> 88  05c7067c4f4fe4fcb5fad606097815c1
#> 89  05c7067c4f4fe4fcb5fad606097815c1
#> 90  7e3aefa426aa39c40a63405b6dfec950
#> 91  05c7067c4f4fe4fcb5fad606097815c1
#> 92  05c7067c4f4fe4fcb5fad606097815c1
#> 93  497ae8d6e1657de7bfbb995b3d0555d0
#> 94  497ae8d6e1657de7bfbb995b3d0555d0
#> 95  497ae8d6e1657de7bfbb995b3d0555d0
#> 96  497ae8d6e1657de7bfbb995b3d0555d0
#> 97  497ae8d6e1657de7bfbb995b3d0555d0
#> 98  497ae8d6e1657de7bfbb995b3d0555d0
#> 99  497ae8d6e1657de7bfbb995b3d0555d0
#> 100 497ae8d6e1657de7bfbb995b3d0555d0
#> 101 497ae8d6e1657de7bfbb995b3d0555d0
#> 102 497ae8d6e1657de7bfbb995b3d0555d0
#> 103 497ae8d6e1657de7bfbb995b3d0555d0
#> 104 497ae8d6e1657de7bfbb995b3d0555d0
#> 105 497ae8d6e1657de7bfbb995b3d0555d0
#> 106 7e3aefa426aa39c40a63405b6dfec950
#> 107 497ae8d6e1657de7bfbb995b3d0555d0
#> 108 497ae8d6e1657de7bfbb995b3d0555d0
#> 109 d6c834d34a9be232e2ada1455143cac6
#> 110 d6c834d34a9be232e2ada1455143cac6
#> 111 d6c834d34a9be232e2ada1455143cac6
#> 112 d6c834d34a9be232e2ada1455143cac6
#> 113 d6c834d34a9be232e2ada1455143cac6
#> 114 d6c834d34a9be232e2ada1455143cac6
#> 115 d6c834d34a9be232e2ada1455143cac6
#> 116 d6c834d34a9be232e2ada1455143cac6
#> 117 d6c834d34a9be232e2ada1455143cac6
#> 118 d6c834d34a9be232e2ada1455143cac6
#> 119 d6c834d34a9be232e2ada1455143cac6
#> 120 7e3aefa426aa39c40a63405b6dfec950
#> 121 d6c834d34a9be232e2ada1455143cac6
#> 122 d6c834d34a9be232e2ada1455143cac6
#> 123 d87c917f988f8035157eeac8c1acb1a3
#> 124 d87c917f988f8035157eeac8c1acb1a3
#> 125 d87c917f988f8035157eeac8c1acb1a3
#> 126 d87c917f988f8035157eeac8c1acb1a3
#> 127 d87c917f988f8035157eeac8c1acb1a3
#> 128 d87c917f988f8035157eeac8c1acb1a3
#> 129 d87c917f988f8035157eeac8c1acb1a3
#> 130 d87c917f988f8035157eeac8c1acb1a3
#> 131 d87c917f988f8035157eeac8c1acb1a3
#> 132 d87c917f988f8035157eeac8c1acb1a3
#> 133 d87c917f988f8035157eeac8c1acb1a3
#> 134 d87c917f988f8035157eeac8c1acb1a3
#> 135 d87c917f988f8035157eeac8c1acb1a3
#> 136 7e3aefa426aa39c40a63405b6dfec950
#> 137 d87c917f988f8035157eeac8c1acb1a3
#> 138 d87c917f988f8035157eeac8c1acb1a3
#> 139 d87c917f988f8035157eeac8c1acb1a3
#> 140 d87c917f988f8035157eeac8c1acb1a3
#> 141 d87c917f988f8035157eeac8c1acb1a3
#> 142 d87c917f988f8035157eeac8c1acb1a3
#> 143 d87c917f988f8035157eeac8c1acb1a3
#> 144 d87c917f988f8035157eeac8c1acb1a3
#> 145 d87c917f988f8035157eeac8c1acb1a3
#> 146 d87c917f988f8035157eeac8c1acb1a3
#> 147 d87c917f988f8035157eeac8c1acb1a3
#> 148 d87c917f988f8035157eeac8c1acb1a3
#> 149 d87c917f988f8035157eeac8c1acb1a3
#> 150 7e3aefa426aa39c40a63405b6dfec950
#> 151 d87c917f988f8035157eeac8c1acb1a3
#> 152 d87c917f988f8035157eeac8c1acb1a3
#> 153 d87c917f988f8035157eeac8c1acb1a3
#> 154 d87c917f988f8035157eeac8c1acb1a3
#> 155 d87c917f988f8035157eeac8c1acb1a3
#> 156 d87c917f988f8035157eeac8c1acb1a3
#> 157 d87c917f988f8035157eeac8c1acb1a3
#> 158 d87c917f988f8035157eeac8c1acb1a3
#> 159 d87c917f988f8035157eeac8c1acb1a3
#> 160 d87c917f988f8035157eeac8c1acb1a3
#> 161 d87c917f988f8035157eeac8c1acb1a3
#> 162 d87c917f988f8035157eeac8c1acb1a3
#> 163 d87c917f988f8035157eeac8c1acb1a3
#> 164 7e3aefa426aa39c40a63405b6dfec950
#> 165 d87c917f988f8035157eeac8c1acb1a3
#> 166 d87c917f988f8035157eeac8c1acb1a3
#> 167 61151f9db7da9f6e15f485cf083d4f89
#> 168 61151f9db7da9f6e15f485cf083d4f89
#> 169 61151f9db7da9f6e15f485cf083d4f89
#> 170 61151f9db7da9f6e15f485cf083d4f89
#> 171 61151f9db7da9f6e15f485cf083d4f89
#> 172 61151f9db7da9f6e15f485cf083d4f89
#> 173 7e3aefa426aa39c40a63405b6dfec950
#> 174 61151f9db7da9f6e15f485cf083d4f89
#>                                                                                                                                                    tag
#> 1                                                                                                                                           format:rda
#> 2                                                                                                                                          name:mtcars
#> 3                                                                                                                                     class:data.frame
#> 4                                                                                                                                          varname:mpg
#> 5                                                                                                                                          varname:cyl
#> 6                                                                                                                                         varname:disp
#> 7                                                                                                                                           varname:hp
#> 8                                                                                                                                         varname:drat
#> 9                                                                                                                                           varname:wt
#> 10                                                                                                                                        varname:qsec
#> 11                                                                                                                                          varname:vs
#> 12                                                                                                                                          varname:am
#> 13                                                                                                                                        varname:gear
#> 14                                                                                                                                        varname:carb
#> 15                                                                                                                            date:2017-11-25 01:47:16
#> 16                                                                                                                                          format:rda
#> 17                                                                                                       session_info:7e3aefa426aa39c40a63405b6dfec950
#> 18                                                                                                                                          format:txt
#> 19                                                                                                                                          format:rda
#> 20                                                                                                                                        name:res_val
#> 21                                                                                                                                    class:grouped_df
#> 22                                                                                                                                        class:tbl_df
#> 23                                                                                                                                           class:tbl
#> 24                                                                                                                                    class:data.frame
#> 25                                                                                                                                         varname:mpg
#> 26                                                                                                                                         varname:cyl
#> 27                                                                                                                                        varname:disp
#> 28                                                                                                                                          varname:hp
#> 29                                                                                                                                        varname:drat
#> 30                                                                                                                                          varname:wt
#> 31                                                                                                                                        varname:qsec
#> 32                                                                                                                                          varname:vs
#> 33                                                                                                                                          varname:am
#> 34                                                                                                                                        varname:gear
#> 35                                                                                                                                        varname:carb
#> 36                                                                                                                            date:2017-11-25 01:47:16
#> 37                                                                                                                LHS:a63c70e73b58d0823ab3bcbd3b543d6f
#> 38                                                                                                                               RHS:group_by(cyl, am)
#> 39                                                                                                                                          format:rda
#> 40                                                                                                       session_info:7e3aefa426aa39c40a63405b6dfec950
#> 41                                                                                                                                          format:txt
#> 42                                                                                                                                          format:rda
#> 43                                                                                                                   name:mtcars %a% group_by(cyl, am)
#> 44                                                                                                                                    class:grouped_df
#> 45                                                                                                                                        class:tbl_df
#> 46                                                                                                                                           class:tbl
#> 47                                                                                                                                    class:data.frame
#> 48                                                                                                                                         varname:mpg
#> 49                                                                                                                                         varname:cyl
#> 50                                                                                                                                        varname:disp
#> 51                                                                                                                                          varname:hp
#> 52                                                                                                                                        varname:drat
#> 53                                                                                                                                          varname:wt
#> 54                                                                                                                                        varname:qsec
#> 55                                                                                                                                          varname:vs
#> 56                                                                                                                                          varname:am
#> 57                                                                                                                                        varname:gear
#> 58                                                                                                                                        varname:carb
#> 59                                                                                                                            date:2017-11-25 01:47:16
#> 60                                                                                                                                          format:rda
#> 61                                                                                                       session_info:7e3aefa426aa39c40a63405b6dfec950
#> 62                                                                                                                                          format:txt
#> 63                                                                                                                                          format:rda
#> 64                                                                                                                                        name:res_val
#> 65                                                                                                                                    class:grouped_df
#> 66                                                                                                                                        class:tbl_df
#> 67                                                                                                                                           class:tbl
#> 68                                                                                                                                    class:data.frame
#> 69                                                                                                                                         varname:mpg
#> 70                                                                                                                                         varname:cyl
#> 71                                                                                                                                          varname:wt
#> 72                                                                                                                                          varname:am
#> 73                                                                                                                            date:2017-11-25 01:47:16
#> 74                                                                                                                LHS:403c3921f0fb68cbc8d5c6ca18e49141
#> 75                                                                                                                        RHS:select(mpg, cyl, wt, am)
#> 76                                                                                                                                          format:rda
#> 77                                                                                                       session_info:7e3aefa426aa39c40a63405b6dfec950
#> 78                                                                                                                                          format:txt
#> 79                                                                                                                                          format:rda
#> 80                                                                                      name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am)
#> 81                                                                                                                                    class:grouped_df
#> 82                                                                                                                                        class:tbl_df
#> 83                                                                                                                                           class:tbl
#> 84                                                                                                                                    class:data.frame
#> 85                                                                                                                                         varname:mpg
#> 86                                                                                                                                         varname:cyl
#> 87                                                                                                                                          varname:wt
#> 88                                                                                                                                          varname:am
#> 89                                                                                                                            date:2017-11-25 01:47:17
#> 90                                                                                                                                          format:rda
#> 91                                                                                                       session_info:7e3aefa426aa39c40a63405b6dfec950
#> 92                                                                                                                                          format:txt
#> 93                                                                                                                                          format:rda
#> 94                                                                                                                                        name:res_val
#> 95                                                                                                                                    class:grouped_df
#> 96                                                                                                                                        class:tbl_df
#> 97                                                                                                                                           class:tbl
#> 98                                                                                                                                    class:data.frame
#> 99                                                                                                                                         varname:cyl
#> 100                                                                                                                                         varname:am
#> 101                                                                                                                                     varname:avgmpg
#> 102                                                                                                                                      varname:avgwt
#> 103                                                                                                                           date:2017-11-25 01:47:17
#> 104                                                                                                               LHS:05c7067c4f4fe4fcb5fad606097815c1
#> 105                                                                                                RHS:summarise(avgmpg = mean(mpg), avgwt = mean(wt))
#> 106                                                                                                                                         format:rda
#> 107                                                                                                      session_info:7e3aefa426aa39c40a63405b6dfec950
#> 108                                                                                                                                         format:txt
#> 109                                                                                                                                         format:rda
#> 110                             name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt))
#> 111                                                                                                                                   class:grouped_df
#> 112                                                                                                                                       class:tbl_df
#> 113                                                                                                                                          class:tbl
#> 114                                                                                                                                   class:data.frame
#> 115                                                                                                                                        varname:cyl
#> 116                                                                                                                                         varname:am
#> 117                                                                                                                                     varname:avgmpg
#> 118                                                                                                                                      varname:avgwt
#> 119                                                                                                                           date:2017-11-25 01:47:17
#> 120                                                                                                                                         format:rda
#> 121                                                                                                      session_info:7e3aefa426aa39c40a63405b6dfec950
#> 122                                                                                                                                         format:txt
#> 123                                                                                                                                         format:rda
#> 124                                                                                                                                       name:res_val
#> 125                                                                                                                                   class:grouped_df
#> 126                                                                                                                                       class:tbl_df
#> 127                                                                                                                                          class:tbl
#> 128                                                                                                                                   class:data.frame
#> 129                                                                                                                                        varname:cyl
#> 130                                                                                                                                         varname:am
#> 131                                                                                                                                     varname:avgmpg
#> 132                                                                                                                                      varname:avgwt
#> 133                                                                                                                           date:2017-11-25 01:47:17
#> 134                                                                                                               LHS:d6c834d34a9be232e2ada1455143cac6
#> 135                                                                                                                            RHS:filter(avgmpg > 20)
#> 136                                                                                                                                         format:rda
#> 137                                                                                                      session_info:7e3aefa426aa39c40a63405b6dfec950
#> 138                                                                                                                                         format:txt
#> 139                                                                                                                                         format:rda
#> 140                                                                                                                                         name:__LHS
#> 141                                                                                                                                   class:grouped_df
#> 142                                                                                                                                       class:tbl_df
#> 143                                                                                                                                          class:tbl
#> 144                                                                                                                                   class:data.frame
#> 145                                                                                                                                        varname:cyl
#> 146                                                                                                                                         varname:am
#> 147                                                                                                                                     varname:avgmpg
#> 148                                                                                                                                      varname:avgwt
#> 149                                                                                                                           date:2017-11-25 01:47:17
#> 150                                                                                                                                         format:rda
#> 151                                                                                                      session_info:7e3aefa426aa39c40a63405b6dfec950
#> 152                                                                                                                                         format:txt
#> 153                                                                                                                                         format:rda
#> 154 name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a% filter(avgmpg >     20)
#> 155                                                                                                                                   class:grouped_df
#> 156                                                                                                                                       class:tbl_df
#> 157                                                                                                                                          class:tbl
#> 158                                                                                                                                   class:data.frame
#> 159                                                                                                                                        varname:cyl
#> 160                                                                                                                                         varname:am
#> 161                                                                                                                                     varname:avgmpg
#> 162                                                                                                                                      varname:avgwt
#> 163                                                                                                                           date:2017-11-25 01:47:17
#> 164                                                                                                                                         format:rda
#> 165                                                                                                      session_info:7e3aefa426aa39c40a63405b6dfec950
#> 166                                                                                                                                         format:txt
#> 167                                                                                                                                         format:rda
#> 168                                                                                                                                       name:res_val
#> 169                                                                                                                                    class:character
#> 170                                                                                                                           date:2017-11-25 01:47:17
#> 171                                                                                                               LHS:d87c917f988f8035157eeac8c1acb1a3
#> 172                                                                                                                     RHS:saveToRepo(exampleRepoDir)
#> 173                                                                                                                                         format:rda
#> 174                                                                                                      session_info:7e3aefa426aa39c40a63405b6dfec950
#>             createdDate
#> 1   2017-11-25 01:47:16
#> 2   2017-11-25 01:47:16
#> 3   2017-11-25 01:47:16
#> 4   2017-11-25 01:47:16
#> 5   2017-11-25 01:47:16
#> 6   2017-11-25 01:47:16
#> 7   2017-11-25 01:47:16
#> 8   2017-11-25 01:47:16
#> 9   2017-11-25 01:47:16
#> 10  2017-11-25 01:47:16
#> 11  2017-11-25 01:47:16
#> 12  2017-11-25 01:47:16
#> 13  2017-11-25 01:47:16
#> 14  2017-11-25 01:47:16
#> 15  2017-11-25 01:47:16
#> 16  2017-11-25 01:47:16
#> 17  2017-11-25 01:47:16
#> 18  2017-11-25 01:47:16
#> 19  2017-11-25 01:47:16
#> 20  2017-11-25 01:47:16
#> 21  2017-11-25 01:47:16
#> 22  2017-11-25 01:47:16
#> 23  2017-11-25 01:47:16
#> 24  2017-11-25 01:47:16
#> 25  2017-11-25 01:47:16
#> 26  2017-11-25 01:47:16
#> 27  2017-11-25 01:47:16
#> 28  2017-11-25 01:47:16
#> 29  2017-11-25 01:47:16
#> 30  2017-11-25 01:47:16
#> 31  2017-11-25 01:47:16
#> 32  2017-11-25 01:47:16
#> 33  2017-11-25 01:47:16
#> 34  2017-11-25 01:47:16
#> 35  2017-11-25 01:47:16
#> 36  2017-11-25 01:47:16
#> 37  2017-11-25 01:47:16
#> 38  2017-11-25 01:47:16
#> 39  2017-11-25 01:47:16
#> 40  2017-11-25 01:47:16
#> 41  2017-11-25 01:47:16
#> 42  2017-11-25 01:47:16
#> 43  2017-11-25 01:47:16
#> 44  2017-11-25 01:47:16
#> 45  2017-11-25 01:47:16
#> 46  2017-11-25 01:47:16
#> 47  2017-11-25 01:47:16
#> 48  2017-11-25 01:47:16
#> 49  2017-11-25 01:47:16
#> 50  2017-11-25 01:47:16
#> 51  2017-11-25 01:47:16
#> 52  2017-11-25 01:47:16
#> 53  2017-11-25 01:47:16
#> 54  2017-11-25 01:47:16
#> 55  2017-11-25 01:47:16
#> 56  2017-11-25 01:47:16
#> 57  2017-11-25 01:47:16
#> 58  2017-11-25 01:47:16
#> 59  2017-11-25 01:47:16
#> 60  2017-11-25 01:47:16
#> 61  2017-11-25 01:47:16
#> 62  2017-11-25 01:47:16
#> 63  2017-11-25 01:47:16
#> 64  2017-11-25 01:47:16
#> 65  2017-11-25 01:47:16
#> 66  2017-11-25 01:47:16
#> 67  2017-11-25 01:47:16
#> 68  2017-11-25 01:47:16
#> 69  2017-11-25 01:47:16
#> 70  2017-11-25 01:47:16
#> 71  2017-11-25 01:47:16
#> 72  2017-11-25 01:47:16
#> 73  2017-11-25 01:47:16
#> 74  2017-11-25 01:47:16
#> 75  2017-11-25 01:47:16
#> 76  2017-11-25 01:47:17
#> 77  2017-11-25 01:47:17
#> 78  2017-11-25 01:47:17
#> 79  2017-11-25 01:47:17
#> 80  2017-11-25 01:47:17
#> 81  2017-11-25 01:47:17
#> 82  2017-11-25 01:47:17
#> 83  2017-11-25 01:47:17
#> 84  2017-11-25 01:47:17
#> 85  2017-11-25 01:47:17
#> 86  2017-11-25 01:47:17
#> 87  2017-11-25 01:47:17
#> 88  2017-11-25 01:47:17
#> 89  2017-11-25 01:47:17
#> 90  2017-11-25 01:47:17
#> 91  2017-11-25 01:47:17
#> 92  2017-11-25 01:47:17
#> 93  2017-11-25 01:47:17
#> 94  2017-11-25 01:47:17
#> 95  2017-11-25 01:47:17
#> 96  2017-11-25 01:47:17
#> 97  2017-11-25 01:47:17
#> 98  2017-11-25 01:47:17
#> 99  2017-11-25 01:47:17
#> 100 2017-11-25 01:47:17
#> 101 2017-11-25 01:47:17
#> 102 2017-11-25 01:47:17
#> 103 2017-11-25 01:47:17
#> 104 2017-11-25 01:47:17
#> 105 2017-11-25 01:47:17
#> 106 2017-11-25 01:47:17
#> 107 2017-11-25 01:47:17
#> 108 2017-11-25 01:47:17
#> 109 2017-11-25 01:47:17
#> 110 2017-11-25 01:47:17
#> 111 2017-11-25 01:47:17
#> 112 2017-11-25 01:47:17
#> 113 2017-11-25 01:47:17
#> 114 2017-11-25 01:47:17
#> 115 2017-11-25 01:47:17
#> 116 2017-11-25 01:47:17
#> 117 2017-11-25 01:47:17
#> 118 2017-11-25 01:47:17
#> 119 2017-11-25 01:47:17
#> 120 2017-11-25 01:47:17
#> 121 2017-11-25 01:47:17
#> 122 2017-11-25 01:47:17
#> 123 2017-11-25 01:47:17
#> 124 2017-11-25 01:47:17
#> 125 2017-11-25 01:47:17
#> 126 2017-11-25 01:47:17
#> 127 2017-11-25 01:47:17
#> 128 2017-11-25 01:47:17
#> 129 2017-11-25 01:47:17
#> 130 2017-11-25 01:47:17
#> 131 2017-11-25 01:47:17
#> 132 2017-11-25 01:47:17
#> 133 2017-11-25 01:47:17
#> 134 2017-11-25 01:47:17
#> 135 2017-11-25 01:47:17
#> 136 2017-11-25 01:47:17
#> 137 2017-11-25 01:47:17
#> 138 2017-11-25 01:47:17
#> 139 2017-11-25 01:47:17
#> 140 2017-11-25 01:47:17
#> 141 2017-11-25 01:47:17
#> 142 2017-11-25 01:47:17
#> 143 2017-11-25 01:47:17
#> 144 2017-11-25 01:47:17
#> 145 2017-11-25 01:47:17
#> 146 2017-11-25 01:47:17
#> 147 2017-11-25 01:47:17
#> 148 2017-11-25 01:47:17
#> 149 2017-11-25 01:47:17
#> 150 2017-11-25 01:47:17
#> 151 2017-11-25 01:47:17
#> 152 2017-11-25 01:47:17
#> 153 2017-11-25 01:47:17
#> 154 2017-11-25 01:47:17
#> 155 2017-11-25 01:47:17
#> 156 2017-11-25 01:47:17
#> 157 2017-11-25 01:47:17
#> 158 2017-11-25 01:47:17
#> 159 2017-11-25 01:47:17
#> 160 2017-11-25 01:47:17
#> 161 2017-11-25 01:47:17
#> 162 2017-11-25 01:47:17
#> 163 2017-11-25 01:47:17
#> 164 2017-11-25 01:47:17
#> 165 2017-11-25 01:47:17
#> 166 2017-11-25 01:47:17
#> 167 2017-11-25 01:47:17
#> 168 2017-11-25 01:47:17
#> 169 2017-11-25 01:47:17
#> 170 2017-11-25 01:47:17
#> 171 2017-11-25 01:47:17
#> 172 2017-11-25 01:47:17
#> 173 2017-11-25 01:47:17
#> 174 2017-11-25 01:47:17

# We search for a Tag with default "name" regular expression corresponding to 
# hash md5hash.
getTagsLocal( md5hash = hash, exampleRepoDir )

#> [1] "name:res_val"                                                                                                                                      
#> [2] "name:__LHS"                                                                                                                                        
#> [3] "name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a% filter(avgmpg >     20)"

# Deleting example respository
deleteLocalRepo( exampleRepoDir, TRUE)
rm( exampleRepoDir )

## EXAMPLE with data iris
exampleRepoDir <- tempfile()
createLocalRepo( exampleRepoDir )

data(iris)
saveToRepo(iris, repoDir = exampleRepoDir )

#> [1] "ff575c261c949d073b2895b05d1097c3"

showLocalRepo(exampleRepoDir)

#>                            md5hash                             name
#> 1 ff575c261c949d073b2895b05d1097c3                             iris
#> 2 7e3aefa426aa39c40a63405b6dfec950 7e3aefa426aa39c40a63405b6dfec950
#>           createdDate
#> 1 2017-11-25 01:47:17
#> 2 2017-11-25 01:47:17

showLocalRepo(exampleRepoDir, method = "tags")

#>                            artifact
#> 1  ff575c261c949d073b2895b05d1097c3
#> 2  ff575c261c949d073b2895b05d1097c3
#> 3  ff575c261c949d073b2895b05d1097c3
#> 4  ff575c261c949d073b2895b05d1097c3
#> 5  ff575c261c949d073b2895b05d1097c3
#> 6  ff575c261c949d073b2895b05d1097c3
#> 7  ff575c261c949d073b2895b05d1097c3
#> 8  ff575c261c949d073b2895b05d1097c3
#> 9  ff575c261c949d073b2895b05d1097c3
#> 10 7e3aefa426aa39c40a63405b6dfec950
#> 11 ff575c261c949d073b2895b05d1097c3
#> 12 ff575c261c949d073b2895b05d1097c3
#>                                              tag         createdDate
#> 1                                     format:rda 2017-11-25 01:47:17
#> 2                                      name:iris 2017-11-25 01:47:17
#> 3                               class:data.frame 2017-11-25 01:47:17
#> 4                           varname:Sepal.Length 2017-11-25 01:47:17
#> 5                            varname:Sepal.Width 2017-11-25 01:47:17
#> 6                           varname:Petal.Length 2017-11-25 01:47:17
#> 7                            varname:Petal.Width 2017-11-25 01:47:17
#> 8                                varname:Species 2017-11-25 01:47:17
#> 9                       date:2017-11-25 01:47:17 2017-11-25 01:47:17
#> 10                                    format:rda 2017-11-25 01:47:17
#> 11 session_info:7e3aefa426aa39c40a63405b6dfec950 2017-11-25 01:47:17
#> 12                                    format:txt 2017-11-25 01:47:17

# We can notice that there is only one md5hash 
# (and second for archiveSessionInfo) in repo so we will use it
hash <- showLocalRepo(exampleRepoDir)[1,1]

# We search for a Tag with "varname" regular expression corresponding to 
# hash md5hash.
getTagsLocal( md5hash = hash, exampleRepoDir, tag = "varname" )

#> [1] "varname:Sepal.Length" "varname:Sepal.Width"  "varname:Petal.Length"
#> [4] "varname:Petal.Width"  "varname:Species"     

# There are 5 different Tags with "varname" regular expression

# We needn't use the whole expression "varname". We may use its abbreviation
# and get the same result.
getTagsLocal( md5hash = hash, exampleRepoDir, tag = "varna" )

#> [1] "varname:Sepal.Length" "varname:Sepal.Width"  "varname:Petal.Length"
#> [4] "varname:Petal.Width"  "varname:Species"     

deleteLocalRepo( exampleRepoDir, TRUE)
rm( exampleRepoDir )

### Remote version
## EXAMPLE: pbiecek archivist repository on GitHub

showRemoteRepo(user="pbiecek", repo="archivist")

#>                             md5hash                             name
#> 1  ff575c261c949d073b2895b05d1097c3                             iris
#> 2  f7804cb74efa70d9795ace8c32e333c3                        myplot123
#> 3  5c3b6d82ab29e7b3a2137b6a537e55ec 5c3b6d82ab29e7b3a2137b6a537e55ec
#> 4  cd6557c6163a6f9800f308f343e75e72                            model
#> 5  7a761a2ae54f3d90060a9f6ca04b3506 7a761a2ae54f3d90060a9f6ca04b3506
#> 6  db864b3dcc09888f5f7f2750b4ab553d                             agn1
#> 7  c2380a7b8c090d6564ff12cdd60aaf7f c2380a7b8c090d6564ff12cdd60aaf7f
#> 8  c1ce897de77e8cb793a69b54e7f2b8b5                           fannyx
#> 9  ebe92bb4f9af6c343da4986f84c42b9b ebe92bb4f9af6c343da4986f84c42b9b
#> 10 2ae5f73c0b979532555fb3800458a4d4                             lda1
#> 11 53a9f3aa4235e111524dda17aad2ee3a 53a9f3aa4235e111524dda17aad2ee3a
#> 12 fdf1922d88bb7369896886dfa7da16f6                             qda1
#> 13 4d504b6c315c9be389f2e84186fa64c5 4d504b6c315c9be389f2e84186fa64c5
#> 14 1e05d1a3f103da0d99fba79643f71487                          glmnet1
#> 15 942ca3ecb05524e95375728c114bf619 942ca3ecb05524e95375728c114bf619
#> 16 134ecbbe2a8814d98f0c2758000c408e                     trellis.plot
#> 17 f49104cd336e6417f89141855c0fe4a7 f49104cd336e6417f89141855c0fe4a7
#> 18 df598bf17a3f149e004eede9c68313fa                        this.test
#> 19 3c6ee510336926b494cf8375b903246a 3c6ee510336926b494cf8375b903246a
#> 20 1d85e04b24e5129f0b7577b270d24ab4                            myFit
#> 21 917b56fd9edd81e088f8f7ada48cda95 917b56fd9edd81e088f8f7ada48cda95
#> 22 617ded4953ac986524a1c24703363980 617ded4953ac986524a1c24703363980
#> 23 5629bc43e36d219b613076b17c665eda 5629bc43e36d219b613076b17c665eda
#>            createdDate
#> 1  2015-11-16 12:47:13
#> 2  2015-11-16 12:47:14
#> 3  2015-11-16 12:47:14
#> 4  2015-11-16 12:47:14
#> 5  2015-11-16 12:47:14
#> 6  2015-11-16 12:47:14
#> 7  2015-11-16 12:47:14
#> 8  2015-11-16 12:47:14
#> 9  2015-11-16 12:47:14
#> 10 2015-11-16 12:47:15
#> 11 2015-11-16 12:47:15
#> 12 2015-11-16 12:47:15
#> 13 2015-11-16 12:47:15
#> 14 2015-11-16 12:47:16
#> 15 2015-11-16 12:47:16
#> 16 2015-11-16 12:47:18
#> 17 2015-11-16 12:47:18
#> 18 2015-11-16 12:47:19
#> 19 2015-11-16 12:47:19
#> 20 2015-11-16 12:47:19
#> 21 2015-11-16 12:47:19
#> 22 2015-11-16 12:47:24
#> 23 2015-11-16 12:47:24

# We search for a Tag with default "name" regular expression corresponding to 
# "cd6557c6163a6f9800f308f343e75e72" md5hash.
getTagsRemote( "cd6557c6163a6f9800f308f343e75e72",
                user="pbiecek", repo="archivist")

#> [1] "name:model"

## EXAMPLE: many archivist-like Repositories on one Github repository
# We search for a Tag with default "name" regular expression corresponding to 
# "ff575c261c949d073b2895b05d1097c3" md5hash.
getTagsRemote("ff575c261c949d073b2895b05d1097c3", user="MarcinKosinski",
               repo="Museum", branch="master", subdir="ex1")

#> [1] "name:iris"