md5hash

Repository stores specific values of an artifact, different for various artifact's classes, and artifact themselves. Artifacts are archived with a special attribute named md5hash. To learn more about artifacts visit archivist-package.

Details

For each artifact, md5hash is a unique string of length 32 that is produced by digest function which uses a cryptographical MD5 hash algorithm. The md5hash of each artifact that is archived in the Repository is also saved on the Repository along with the artifact's Tags - see Tags. It enables to distinguish artifacts in the Repository and facilitates searching and loading them.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results URL https://github.com/pbiecek/archivist

See also

Functions that take md5hash as a parameter are:

addTagsRepo,

copyLocalRepo,

copyRemoteRepo,

loadFromLocalRepo,

loadFromRemoteRepo,

getTagsRemote,

getTagsLocal,

rmFromLocalRepo.

Functions returning md5hash as a value are:

saveToLocalRepo,

searchInLocalRepo,

searchInRemoteRepo,

shinySearchInLocalRepo.

Functions returning md5hashes as a data.frame are:

showLocalRepo,

showRemoteRepo.

Learn more about md5hashes at archivist wiki webpage on https://github.com/pbiecek/archivist/wiki/archivist-package-md5hash.

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo