Shiny Based Live Search for an Artifact in a Repository Using Tags

shinySearchInLocalRepo searches for an artifact in a Repository using Tags. To create an application one needs to point the name of artifacts' repository. The application is generated on the run. As for now there are two controllers exposed. A text input field and a slider. Tags that are typed into text field are used for searching in repository. Objects that have the same Tags are presented on the right panel. These object might be also downloaded just by click. To learn more about artifacts visit archivist-package.

shinySearchInLocalRepo(repoDir = NULL, host = "0.0.0.0")

Arguments

repoDir
A character denoting an existing directory in which artifacts will be searched. If set to NULL (by default), uses the repoDir specified in setLocalRepo.

host
A host IP adress, see the host argument in shiny::runApp.

Value

shinySearchInLocalRepo runs a shiny application.

Details

shinySearchInLocalRepo searches for artifacts in a Repository using their Tags (e.g., name, class or archiving date). Tags are submitted in a text input in a shiny application. Many Tags may be specified, they should be comma separated. User can specify more Tags like phase, project, author etc. when artifact is created.

In the search query one can add Tags starting with sort: or sort:-. As a result, miniatures will be sorted appropriately. For example sort:class will sort class Tags, while sort:-class will sort class tags backwards. sort:createdDate will sort createdDate Tag and sort:-createdDate will sort createdDate Tag backwards.

Tags, submitted in the text field, should be given according to the format: "TagKey:TagValue" - see examples.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

shiny

This function use tools from the fantastic shiny package, so you'll need to make sure to have it installed.

References

Biecek P and Kosinski M (2017). "archivist: An R Package for Managing, Recording and Restoring Data Analysis Results." _Journal of Statistical Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL: http://doi.org/10.18637/jss.v082.i11). URL https://github.com/pbiecek/archivist

See also

Examples

# NOT RUN {
  # assuming that there is a 'repo' dir with a valid archivist repository
  shinySearchInLocalRepo( repoDir = 'repo' )
# }