View the List of Artifacts from the Repository

showLocalRepo and showRemoteRepo functions produce the data.frame of the artifacts from the Repository saved in a given repoDir (directory). showLocalRepo shows the artifacts from the Repository that exists on the user's computer whereas showRemoteRepo shows the artifacts of the Repository existing on the remote repository. To learn more about artifacts visit archivist-package.

showLocalRepo(repoDir = aoptions("repoDir"), method = "md5hashes")

showRemoteRepo(repo = aoptions("repo"), user = aoptions("user"),
  branch = aoptions("branch"), subdir = aoptions("subdir"),
  repoType = aoptions("repoType"), method = "md5hashes")

Arguments

repoDir

A character denoting an existing directory of the Repository for which metadata will be returned.

method

A character specifying a method to be used to show the Repository. Available methods: md5hashes (default), tags and sets - see https://github.com/pbiecek/archivist2.

repo

While working with the Remote repository. A character containing a name of the Remote repository on which the Repository is stored. By default set to NULL - see Note.

user

While working with the Remote repository. A character containing a name of the Remote user on whose account the repo is created. By default set to NULL - see Note.

branch

While working with the Remote repository. A character containing a name of the Remote Repository's branch on which the Repository is stored. Default branch is master.

subdir

While working with the Remote repository. A character containing a name of a directory on the Remote repository on which the Repository is stored. If the Repository is stored in the main folder of the Remote repository, this should be set to subdir = "/" as default.

repoType

A character containing a type of the remote repository. Currently it can be 'github' or 'bitbucket'.

Value

If parameter method is set as md5hashes then a data.frame with artifacts' names and artifacts'md5hashes will be returned.

If parameter method is set as tags then a data.frame with Tags and artifacts' md5hashes will be returned.

Also in both cases a data.frame contains an extra column with the date of creation of the Tag or md5hash.

To learn more about Tags or md5hashes check: Tags or md5hash.

Details

showLocalRepo and showRemoteRepo functions produce the data.frame of the artifacts from a Repository saved in a given repoDir (directory). showLocalRepo shows the artifacts from the Repository that exists on the user's computer whereas showRemoteRepo shows the artifacts of the Repository existing on the remote repository.

Both functions show the current state of a Repository, inter alia, all archived artifacts can be seen with their unique md5hash or a data.frame with archived Tags can be obtained.

Note

If repo and user are set to NULL (as default) in the Remote mode then global parameters set in setRemoteRepo (or via aoptions) function are used.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

## Not run: ------------------------------------
# # objects preparation
# 
# showLocalRepo(method = "md5hashes", 
#    repoDir = system.file("graphGallery", package = "archivist"))
# showLocalRepo(method = "tags", 
#    repoDir = system.file("graphGallery", package = "archivist"))
# 
# 
# # Remote version
# 
# showRemoteRepo(method = "md5hashes", user = "pbiecek", repo = "archivist")
# showRemoteRepo(method = "tags", user = "pbiecek", repo = "archivist", branch = "master")
# 
# # many archivist-like Repositories on one Remote repository
# 
# showRemoteRepo( user="MarcinKosinski", repo="Museum", branch="master",
#                 subdir="ex1")
# showRemoteRepo( user="MarcinKosinski", repo="Museum", branch="master",
#                 subdir="ex2")
#                 
# ## Remote options
# showRemoteRepo('archivist', 'pbiecek')
# aoptions('user', 'pbiecek')
# aoptions('repo', 'archivist')
# loadFromRemoteRepo("ff575c261c", value = TRUE) -> iris123
# 
# showRemoteRepo('Museum', 'MarcinKosinski', subdir = 'ex1')
# aoptions('repo', 'Museum')
# aoptions('user', 'MarcinKosinski')
# aoptions('subdir', 'ex1')
# aoptions('branch', 'master')
# showRemoteRepo()
# showRemoteRepo(subdir = 'ex2')
# 
# aoptions('subdir')
# 
# 
## ---------------------------------------------