View the Summary of the Repository

summaryRepo summarizes the current state of the Repository.

summaryLocalRepo(repoDir = aoptions("repoDir"))

summaryRemoteRepo(repo = aoptions("repo"), user = aoptions("user"),
  branch = "master", subdir = aoptions("subdir"),
  repoType = aoptions("repoType"))

Arguments

repoDir
A character denoting an existing directory of the Repository for which a summary will be returned.

repo
While working with the Remote repository. A character containing a name of the Remote repository on which the Repository is stored. By default set to NULL - see Note.

user
While working with the Remote repository. A character containing a name of the Remote user on whose account the repo is created. By default set to NULL - see Note.

branch
While working with the Remote repository. A character containing a name of the Remote Repository's branch on which the Repository is stored. Default branch is master.

subdir
While working with the Remote repository. A character containing a name of a directory on the Remote repository on which the Repository is stored. If the Repository is stored in the main folder of the Remote repository, this should be set to subdir = "/" as default.

repoType
A character containing a type of the remote repository. Currently it can be 'github' or 'bitbucket'.

Value

An object of class repository which can be printed: print(object).

Details

summaryRepo summarizes the current state of a Repository. Recommended to use print( summaryRepo ) ). See examples.

Note

If the same artifact was archived many times then it is counted as one artifact or database in print(summaryRepo).

If repo and user are set to NULL (as default) in the Remote mode then global parameters set in setRemoteRepo function are used.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results Conditionally accepted for Journal of Statistical Software URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, areadLocal, aread, asearchLocal, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, zipLocalRepo

Examples

# NOT RUN {
showLocalRepo(repoDir = system.file("graphGallery", package = "archivist"))
#
# Remote version
#

x <- summaryRemoteRepo( user="pbiecek", repo="archivist")
print( x )

# many archivist-like Repositories on one Remote repository

summaryRemoteRepo(user="MarcinKosinski", repo="Museum",
branch="master", subdir="ex2" )

# }