Create the Summary for Each Artifact in a Markdown Format

createMDGallery creates a summary for each artifact from Repository stored on a GitHub. For each artifact tihd function creates a markdown file with: the download link, artifact's Tags (when addTags = TRUE) and miniature (addMiniature = TRUE) if the artifact was archived with it's miniature and Tags. The miniature is a print or head over an artifact or it's png when it was a plot. But this function only supports png miniatures.

createMDGallery(output, repo = aoptions("repo"),
  user = aoptions("user"), branch = aoptions("branch"),
  subdir = aoptions("subdir"), repoType = aoptions("repoType"),
  addTags = FALSE, addMiniature = FALSE, maxTags = 100)

Arguments

output
A name of the file in which artifacts should be summarized.

repo
A character containing a name of the Remote repository on which the Repository is stored. By default set to NULL - see Note.

user
A character containing a name of the Github user on whose account the repo is created. By default set to NULL - see Note.

branch
A character containing a name of the Remote Repository's branch on which the Repository is stored. Default branch is master.

subdir
A character containing a name of a directory on the Remote repository on which the Repository is stored. If the Repository is stored in the main folder of the Remote repository, this should be set to subdir = "/" as default.

repoType
A character containing a type of the remote repository. Currently it can be 'github' or 'bitbucket'.

addTags
Logical, whether to add artfiact's Tags to the output.

addMiniature
Logical, whether to add artfiact's miniature/plots to the output.

maxTags
Integer. The maximal length of chunks output when describing Tags of artifact.

Details

To learn more about artifacts visit archivist-package.

Note

If repo and user are set to NULL (as default) in the Remote mode then global parameters set in setRemoteRepo (or via aoptions) function are used.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Biecek P and Kosinski M (2017). "archivist: An R Package for Managing, Recording and Restoring Data Analysis Results." _Journal of Statistical Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL: http://doi.org/10.18637/jss.v082.i11). URL https://github.com/pbiecek/archivist

See also

Examples

# NOT RUN {
createMDGallery(user = 'MarcinKosinski', repo = 'Museum',
 'README_test1.md', addTags = TRUE)
createMDGallery('graphGallery', 'pbiecek', addMiniature = TRUE,
 'README_test2.md', addTags = TRUE)

# }