Delete the Existing Repository from the Given Directory

deleteLocalRepo deletes the existing Repository from the given directory. As a result all artifacts from gallery folder are removed and database backpack.db is deleted.

deleteLocalRepo(repoDir, deleteRoot = FALSE, unset = FALSE)

deleteRepo(...)

Arguments

repoDir
A character that specifies the directory for the Repository which is to be deleted.

deleteRoot
A logical value that specifies if the repository root directory should be deleted for Local Repository.

unset
A logical. If deleted repoDir was set to be default Local Repository and unset is TRUE, then repoDir is unset as a default Local Repository (aoptions('repoDir/repo', NULL, T)).

...
All arguments are being passed to deleteLocalRepo.

Note

Remember that using tempfile() instead of tempdir() in examples section is crucial. tempdir() is existing directory in which R works so calling deleteLocalRepo(exampleRepoDir, deleteRoot=TRUE) removes important R files. You can find out more information about this problem at stackoverflow webpage.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results Conditionally accepted for Journal of Statistical Software URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

exampleRepoDir <- tempfile()
createLocalRepo( repoDir = exampleRepoDir, default =  TRUE )

#> Directory /var/folders/_l/jllqh32s3llbxmtrh2431vpr0000gn/T//RtmpDvQURu/file107d6627d0fac did not exist. Forced to create a new directory.

data(iris)
saveToLocalRepo(iris)

#> [1] "ff575c261c949d073b2895b05d1097c3"

deleteLocalRepo( repoDir = exampleRepoDir, unset = TRUE, deleteRoot = TRUE)

#> NULL