Return Tags Corresponding to md5hash

getTagsLocal and getTagsRemote return Tags (see Tags) related to md5hash of an artifact. To learn more about artifacts visit archivist-package.

getTagsLocal(md5hash, repoDir = aoptions("repoDir"), tag = "name")

getTagsRemote(md5hash, repo = aoptions("repo"), user = aoptions("user"),
  branch = aoptions("branch"), subdir = aoptions("subdir"),
  repoType = aoptions("repoType"), tag = "name")

Arguments

md5hash

A character containing md5hash of artifacts which Tags are desired to be returned.

repoDir

A character denoting an existing directory in which artifacts are stored.

tag

A regular expression denoting type of a Tag that we search for (see Examples). Default tag = "name".

repo

While working with the Remote repository. A character containing a name of the Remote repository on which the Repository is stored. By default set to NULL - see Note.

user

While working with the Remote repository. A character containing a name of the Remote user on whose account the repo is created. By default set to NULL - see Note.

branch

While working with the Remote repository. A character containing a name of the Remote repository's branch on which the Repository is stored. Default branch is master.

subdir

While working with the Remote repository. A character containing a name of a directory on the Remote repository on which the Repository is stored. If the Repository is stored in main folder on the Remote repository, this should be set to subdir = "/" as default.

repoType

A character containing a type of the remote repository. Currently it can be 'github' or 'bitbucket'.

Value

The character vector of Tags (see Tags) related to md5hash of an artifact.

Details

getTagsLocal and getTagsRemote return Tags, of a specific type described by tag parameter, related to md5hash of an artifact. To learn more about artifacts visit archivist-package.

Note

If repo and user are set to NULL (as default) in Remote mode then global parameters set in setRemoteRepo function are used.

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Przemyslaw Biecek, Marcin Kosinski (2017). archivist: An R Package for Managing, Recording and Restoring Data Analysis Results URL https://github.com/pbiecek/archivist

See also

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, loadFromLocalRepo, md5hash, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, aread, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, loadFromLocalRepo, md5hash, restoreLibs, rmFromLocalRepo, saveToLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

### Local version

## EXAMPLE with pipe operator %a%

# Creating empty repository
exampleRepoDir <- tempfile()
createLocalRepo( exampleRepoDir )

#> Directory /tmp/RtmpFa89Nv/file31a7fc19b4f did not exist. Forced to create a new directory.

library(dplyr)
data(mtcars)
setLocalRepo(repoDir = exampleRepoDir)
hash <- mtcars %a%
 group_by(cyl, am) %a%
   select(mpg, cyl, wt, am) %a%
     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a%
       filter(avgmpg > 20) %a%
       saveToRepo( exampleRepoDir )

showLocalRepo(exampleRepoDir)

#>                             md5hash
#> 1  a63c70e73b58d0823ab3bcbd3b543d6f
#> 2  2cb4e61dcc8f3c8bb347807aa59461ca
#> 3  0de93da714e397a9640cb5bac6d85ecf
#> 4  2cb4e61dcc8f3c8bb347807aa59461ca
#> 5  0de93da714e397a9640cb5bac6d85ecf
#> 6  2cb4e61dcc8f3c8bb347807aa59461ca
#> 7  0794e49145b4a2467a9986026c710498
#> 8  2cb4e61dcc8f3c8bb347807aa59461ca
#> 9  0794e49145b4a2467a9986026c710498
#> 10 2cb4e61dcc8f3c8bb347807aa59461ca
#> 11 c3be68c919026e0876e4ba8b783945db
#> 12 2cb4e61dcc8f3c8bb347807aa59461ca
#> 13 40554c7e8cff1adbacf3405624fd58ce
#> 14 2cb4e61dcc8f3c8bb347807aa59461ca
#> 15 177cbfc0062d297ae58cd38566014ba8
#> 16 2cb4e61dcc8f3c8bb347807aa59461ca
#> 17 177cbfc0062d297ae58cd38566014ba8
#> 18 2cb4e61dcc8f3c8bb347807aa59461ca
#> 19 177cbfc0062d297ae58cd38566014ba8
#> 20 2cb4e61dcc8f3c8bb347807aa59461ca
#> 21 6d3b822cfee0fa27333a854c10a4db4c
#> 22 2cb4e61dcc8f3c8bb347807aa59461ca
#>                                                                                                                                             name
#> 1                                                                                                                                         mtcars
#> 2                                                                                                               2cb4e61dcc8f3c8bb347807aa59461ca
#> 3                                                                                                                                        res_val
#> 4                                                                                                               2cb4e61dcc8f3c8bb347807aa59461ca
#> 5                                                                                                                   mtcars %a% group_by(cyl, am)
#> 6                                                                                                               2cb4e61dcc8f3c8bb347807aa59461ca
#> 7                                                                                                                                        res_val
#> 8                                                                                                               2cb4e61dcc8f3c8bb347807aa59461ca
#> 9                                                                                      mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am)
#> 10                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#> 11                                                                                                                                       res_val
#> 12                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#> 13                             mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt))
#> 14                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#> 15                                                                                                                                       res_val
#> 16                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#> 17                                                                                                                                         __LHS
#> 18                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#> 19 mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a% filter(avgmpg >     20)
#> 20                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#> 21                                                                                                                                       res_val
#> 22                                                                                                              2cb4e61dcc8f3c8bb347807aa59461ca
#>            createdDate
#> 1  2017-03-15 16:40:31
#> 2  2017-03-15 16:40:31
#> 3  2017-03-15 16:40:31
#> 4  2017-03-15 16:40:31
#> 5  2017-03-15 16:40:31
#> 6  2017-03-15 16:40:31
#> 7  2017-03-15 16:40:31
#> 8  2017-03-15 16:40:31
#> 9  2017-03-15 16:40:31
#> 10 2017-03-15 16:40:31
#> 11 2017-03-15 16:40:31
#> 12 2017-03-15 16:40:31
#> 13 2017-03-15 16:40:31
#> 14 2017-03-15 16:40:31
#> 15 2017-03-15 16:40:31
#> 16 2017-03-15 16:40:31
#> 17 2017-03-15 16:40:31
#> 18 2017-03-15 16:40:31
#> 19 2017-03-15 16:40:31
#> 20 2017-03-15 16:40:31
#> 21 2017-03-15 16:40:31
#> 22 2017-03-15 16:40:31

showLocalRepo(exampleRepoDir, method = "tags")

#>                             artifact
#> 1   a63c70e73b58d0823ab3bcbd3b543d6f
#> 2   a63c70e73b58d0823ab3bcbd3b543d6f
#> 3   a63c70e73b58d0823ab3bcbd3b543d6f
#> 4   a63c70e73b58d0823ab3bcbd3b543d6f
#> 5   a63c70e73b58d0823ab3bcbd3b543d6f
#> 6   a63c70e73b58d0823ab3bcbd3b543d6f
#> 7   a63c70e73b58d0823ab3bcbd3b543d6f
#> 8   a63c70e73b58d0823ab3bcbd3b543d6f
#> 9   a63c70e73b58d0823ab3bcbd3b543d6f
#> 10  a63c70e73b58d0823ab3bcbd3b543d6f
#> 11  a63c70e73b58d0823ab3bcbd3b543d6f
#> 12  a63c70e73b58d0823ab3bcbd3b543d6f
#> 13  a63c70e73b58d0823ab3bcbd3b543d6f
#> 14  a63c70e73b58d0823ab3bcbd3b543d6f
#> 15  a63c70e73b58d0823ab3bcbd3b543d6f
#> 16  2cb4e61dcc8f3c8bb347807aa59461ca
#> 17  a63c70e73b58d0823ab3bcbd3b543d6f
#> 18  a63c70e73b58d0823ab3bcbd3b543d6f
#> 19  0de93da714e397a9640cb5bac6d85ecf
#> 20  0de93da714e397a9640cb5bac6d85ecf
#> 21  0de93da714e397a9640cb5bac6d85ecf
#> 22  0de93da714e397a9640cb5bac6d85ecf
#> 23  0de93da714e397a9640cb5bac6d85ecf
#> 24  0de93da714e397a9640cb5bac6d85ecf
#> 25  0de93da714e397a9640cb5bac6d85ecf
#> 26  0de93da714e397a9640cb5bac6d85ecf
#> 27  0de93da714e397a9640cb5bac6d85ecf
#> 28  0de93da714e397a9640cb5bac6d85ecf
#> 29  0de93da714e397a9640cb5bac6d85ecf
#> 30  0de93da714e397a9640cb5bac6d85ecf
#> 31  0de93da714e397a9640cb5bac6d85ecf
#> 32  0de93da714e397a9640cb5bac6d85ecf
#> 33  0de93da714e397a9640cb5bac6d85ecf
#> 34  0de93da714e397a9640cb5bac6d85ecf
#> 35  0de93da714e397a9640cb5bac6d85ecf
#> 36  0de93da714e397a9640cb5bac6d85ecf
#> 37  0de93da714e397a9640cb5bac6d85ecf
#> 38  0de93da714e397a9640cb5bac6d85ecf
#> 39  2cb4e61dcc8f3c8bb347807aa59461ca
#> 40  0de93da714e397a9640cb5bac6d85ecf
#> 41  0de93da714e397a9640cb5bac6d85ecf
#> 42  0de93da714e397a9640cb5bac6d85ecf
#> 43  0de93da714e397a9640cb5bac6d85ecf
#> 44  0de93da714e397a9640cb5bac6d85ecf
#> 45  0de93da714e397a9640cb5bac6d85ecf
#> 46  0de93da714e397a9640cb5bac6d85ecf
#> 47  0de93da714e397a9640cb5bac6d85ecf
#> 48  0de93da714e397a9640cb5bac6d85ecf
#> 49  0de93da714e397a9640cb5bac6d85ecf
#> 50  0de93da714e397a9640cb5bac6d85ecf
#> 51  0de93da714e397a9640cb5bac6d85ecf
#> 52  0de93da714e397a9640cb5bac6d85ecf
#> 53  0de93da714e397a9640cb5bac6d85ecf
#> 54  0de93da714e397a9640cb5bac6d85ecf
#> 55  0de93da714e397a9640cb5bac6d85ecf
#> 56  0de93da714e397a9640cb5bac6d85ecf
#> 57  0de93da714e397a9640cb5bac6d85ecf
#> 58  0de93da714e397a9640cb5bac6d85ecf
#> 59  0de93da714e397a9640cb5bac6d85ecf
#> 60  2cb4e61dcc8f3c8bb347807aa59461ca
#> 61  0de93da714e397a9640cb5bac6d85ecf
#> 62  0de93da714e397a9640cb5bac6d85ecf
#> 63  0794e49145b4a2467a9986026c710498
#> 64  0794e49145b4a2467a9986026c710498
#> 65  0794e49145b4a2467a9986026c710498
#> 66  0794e49145b4a2467a9986026c710498
#> 67  0794e49145b4a2467a9986026c710498
#> 68  0794e49145b4a2467a9986026c710498
#> 69  0794e49145b4a2467a9986026c710498
#> 70  0794e49145b4a2467a9986026c710498
#> 71  0794e49145b4a2467a9986026c710498
#> 72  0794e49145b4a2467a9986026c710498
#> 73  0794e49145b4a2467a9986026c710498
#> 74  0794e49145b4a2467a9986026c710498
#> 75  0794e49145b4a2467a9986026c710498
#> 76  2cb4e61dcc8f3c8bb347807aa59461ca
#> 77  0794e49145b4a2467a9986026c710498
#> 78  0794e49145b4a2467a9986026c710498
#> 79  0794e49145b4a2467a9986026c710498
#> 80  0794e49145b4a2467a9986026c710498
#> 81  0794e49145b4a2467a9986026c710498
#> 82  0794e49145b4a2467a9986026c710498
#> 83  0794e49145b4a2467a9986026c710498
#> 84  0794e49145b4a2467a9986026c710498
#> 85  0794e49145b4a2467a9986026c710498
#> 86  0794e49145b4a2467a9986026c710498
#> 87  0794e49145b4a2467a9986026c710498
#> 88  0794e49145b4a2467a9986026c710498
#> 89  0794e49145b4a2467a9986026c710498
#> 90  2cb4e61dcc8f3c8bb347807aa59461ca
#> 91  0794e49145b4a2467a9986026c710498
#> 92  0794e49145b4a2467a9986026c710498
#> 93  c3be68c919026e0876e4ba8b783945db
#> 94  c3be68c919026e0876e4ba8b783945db
#> 95  c3be68c919026e0876e4ba8b783945db
#> 96  c3be68c919026e0876e4ba8b783945db
#> 97  c3be68c919026e0876e4ba8b783945db
#> 98  c3be68c919026e0876e4ba8b783945db
#> 99  c3be68c919026e0876e4ba8b783945db
#> 100 c3be68c919026e0876e4ba8b783945db
#> 101 c3be68c919026e0876e4ba8b783945db
#> 102 c3be68c919026e0876e4ba8b783945db
#> 103 c3be68c919026e0876e4ba8b783945db
#> 104 c3be68c919026e0876e4ba8b783945db
#> 105 c3be68c919026e0876e4ba8b783945db
#> 106 2cb4e61dcc8f3c8bb347807aa59461ca
#> 107 c3be68c919026e0876e4ba8b783945db
#> 108 c3be68c919026e0876e4ba8b783945db
#> 109 40554c7e8cff1adbacf3405624fd58ce
#> 110 40554c7e8cff1adbacf3405624fd58ce
#> 111 40554c7e8cff1adbacf3405624fd58ce
#> 112 40554c7e8cff1adbacf3405624fd58ce
#> 113 40554c7e8cff1adbacf3405624fd58ce
#> 114 40554c7e8cff1adbacf3405624fd58ce
#> 115 40554c7e8cff1adbacf3405624fd58ce
#> 116 40554c7e8cff1adbacf3405624fd58ce
#> 117 40554c7e8cff1adbacf3405624fd58ce
#> 118 40554c7e8cff1adbacf3405624fd58ce
#> 119 40554c7e8cff1adbacf3405624fd58ce
#> 120 2cb4e61dcc8f3c8bb347807aa59461ca
#> 121 40554c7e8cff1adbacf3405624fd58ce
#> 122 40554c7e8cff1adbacf3405624fd58ce
#> 123 177cbfc0062d297ae58cd38566014ba8
#> 124 177cbfc0062d297ae58cd38566014ba8
#> 125 177cbfc0062d297ae58cd38566014ba8
#> 126 177cbfc0062d297ae58cd38566014ba8
#> 127 177cbfc0062d297ae58cd38566014ba8
#> 128 177cbfc0062d297ae58cd38566014ba8
#> 129 177cbfc0062d297ae58cd38566014ba8
#> 130 177cbfc0062d297ae58cd38566014ba8
#> 131 177cbfc0062d297ae58cd38566014ba8
#> 132 177cbfc0062d297ae58cd38566014ba8
#> 133 177cbfc0062d297ae58cd38566014ba8
#> 134 177cbfc0062d297ae58cd38566014ba8
#> 135 177cbfc0062d297ae58cd38566014ba8
#> 136 2cb4e61dcc8f3c8bb347807aa59461ca
#> 137 177cbfc0062d297ae58cd38566014ba8
#> 138 177cbfc0062d297ae58cd38566014ba8
#> 139 177cbfc0062d297ae58cd38566014ba8
#> 140 177cbfc0062d297ae58cd38566014ba8
#> 141 177cbfc0062d297ae58cd38566014ba8
#> 142 177cbfc0062d297ae58cd38566014ba8
#> 143 177cbfc0062d297ae58cd38566014ba8
#> 144 177cbfc0062d297ae58cd38566014ba8
#> 145 177cbfc0062d297ae58cd38566014ba8
#> 146 177cbfc0062d297ae58cd38566014ba8
#> 147 177cbfc0062d297ae58cd38566014ba8
#> 148 177cbfc0062d297ae58cd38566014ba8
#> 149 177cbfc0062d297ae58cd38566014ba8
#> 150 2cb4e61dcc8f3c8bb347807aa59461ca
#> 151 177cbfc0062d297ae58cd38566014ba8
#> 152 177cbfc0062d297ae58cd38566014ba8
#> 153 177cbfc0062d297ae58cd38566014ba8
#> 154 177cbfc0062d297ae58cd38566014ba8
#> 155 177cbfc0062d297ae58cd38566014ba8
#> 156 177cbfc0062d297ae58cd38566014ba8
#> 157 177cbfc0062d297ae58cd38566014ba8
#> 158 177cbfc0062d297ae58cd38566014ba8
#> 159 177cbfc0062d297ae58cd38566014ba8
#> 160 177cbfc0062d297ae58cd38566014ba8
#> 161 177cbfc0062d297ae58cd38566014ba8
#> 162 177cbfc0062d297ae58cd38566014ba8
#> 163 177cbfc0062d297ae58cd38566014ba8
#> 164 2cb4e61dcc8f3c8bb347807aa59461ca
#> 165 177cbfc0062d297ae58cd38566014ba8
#> 166 177cbfc0062d297ae58cd38566014ba8
#> 167 6d3b822cfee0fa27333a854c10a4db4c
#> 168 6d3b822cfee0fa27333a854c10a4db4c
#> 169 6d3b822cfee0fa27333a854c10a4db4c
#> 170 6d3b822cfee0fa27333a854c10a4db4c
#> 171 6d3b822cfee0fa27333a854c10a4db4c
#> 172 6d3b822cfee0fa27333a854c10a4db4c
#> 173 2cb4e61dcc8f3c8bb347807aa59461ca
#> 174 6d3b822cfee0fa27333a854c10a4db4c
#>                                                                                                                                                    tag
#> 1                                                                                                                                           format:rda
#> 2                                                                                                                                          name:mtcars
#> 3                                                                                                                                     class:data.frame
#> 4                                                                                                                                          varname:mpg
#> 5                                                                                                                                          varname:cyl
#> 6                                                                                                                                         varname:disp
#> 7                                                                                                                                           varname:hp
#> 8                                                                                                                                         varname:drat
#> 9                                                                                                                                           varname:wt
#> 10                                                                                                                                        varname:qsec
#> 11                                                                                                                                          varname:vs
#> 12                                                                                                                                          varname:am
#> 13                                                                                                                                        varname:gear
#> 14                                                                                                                                        varname:carb
#> 15                                                                                                                            date:2017-03-15 16:40:31
#> 16                                                                                                                                          format:rda
#> 17                                                                                                       session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 18                                                                                                                                          format:txt
#> 19                                                                                                                                          format:rda
#> 20                                                                                                                                        name:res_val
#> 21                                                                                                                                    class:grouped_df
#> 22                                                                                                                                        class:tbl_df
#> 23                                                                                                                                           class:tbl
#> 24                                                                                                                                    class:data.frame
#> 25                                                                                                                                         varname:mpg
#> 26                                                                                                                                         varname:cyl
#> 27                                                                                                                                        varname:disp
#> 28                                                                                                                                          varname:hp
#> 29                                                                                                                                        varname:drat
#> 30                                                                                                                                          varname:wt
#> 31                                                                                                                                        varname:qsec
#> 32                                                                                                                                          varname:vs
#> 33                                                                                                                                          varname:am
#> 34                                                                                                                                        varname:gear
#> 35                                                                                                                                        varname:carb
#> 36                                                                                                                            date:2017-03-15 16:40:31
#> 37                                                                                                                LHS:a63c70e73b58d0823ab3bcbd3b543d6f
#> 38                                                                                                                               RHS:group_by(cyl, am)
#> 39                                                                                                                                          format:rda
#> 40                                                                                                       session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 41                                                                                                                                          format:txt
#> 42                                                                                                                                          format:rda
#> 43                                                                                                                   name:mtcars %a% group_by(cyl, am)
#> 44                                                                                                                                    class:grouped_df
#> 45                                                                                                                                        class:tbl_df
#> 46                                                                                                                                           class:tbl
#> 47                                                                                                                                    class:data.frame
#> 48                                                                                                                                         varname:mpg
#> 49                                                                                                                                         varname:cyl
#> 50                                                                                                                                        varname:disp
#> 51                                                                                                                                          varname:hp
#> 52                                                                                                                                        varname:drat
#> 53                                                                                                                                          varname:wt
#> 54                                                                                                                                        varname:qsec
#> 55                                                                                                                                          varname:vs
#> 56                                                                                                                                          varname:am
#> 57                                                                                                                                        varname:gear
#> 58                                                                                                                                        varname:carb
#> 59                                                                                                                            date:2017-03-15 16:40:31
#> 60                                                                                                                                          format:rda
#> 61                                                                                                       session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 62                                                                                                                                          format:txt
#> 63                                                                                                                                          format:rda
#> 64                                                                                                                                        name:res_val
#> 65                                                                                                                                    class:grouped_df
#> 66                                                                                                                                        class:tbl_df
#> 67                                                                                                                                           class:tbl
#> 68                                                                                                                                    class:data.frame
#> 69                                                                                                                                         varname:mpg
#> 70                                                                                                                                         varname:cyl
#> 71                                                                                                                                          varname:wt
#> 72                                                                                                                                          varname:am
#> 73                                                                                                                            date:2017-03-15 16:40:31
#> 74                                                                                                                LHS:0de93da714e397a9640cb5bac6d85ecf
#> 75                                                                                                                        RHS:select(mpg, cyl, wt, am)
#> 76                                                                                                                                          format:rda
#> 77                                                                                                       session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 78                                                                                                                                          format:txt
#> 79                                                                                                                                          format:rda
#> 80                                                                                      name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am)
#> 81                                                                                                                                    class:grouped_df
#> 82                                                                                                                                        class:tbl_df
#> 83                                                                                                                                           class:tbl
#> 84                                                                                                                                    class:data.frame
#> 85                                                                                                                                         varname:mpg
#> 86                                                                                                                                         varname:cyl
#> 87                                                                                                                                          varname:wt
#> 88                                                                                                                                          varname:am
#> 89                                                                                                                            date:2017-03-15 16:40:31
#> 90                                                                                                                                          format:rda
#> 91                                                                                                       session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 92                                                                                                                                          format:txt
#> 93                                                                                                                                          format:rda
#> 94                                                                                                                                        name:res_val
#> 95                                                                                                                                    class:grouped_df
#> 96                                                                                                                                        class:tbl_df
#> 97                                                                                                                                           class:tbl
#> 98                                                                                                                                    class:data.frame
#> 99                                                                                                                                         varname:cyl
#> 100                                                                                                                                         varname:am
#> 101                                                                                                                                     varname:avgmpg
#> 102                                                                                                                                      varname:avgwt
#> 103                                                                                                                           date:2017-03-15 16:40:31
#> 104                                                                                                               LHS:0794e49145b4a2467a9986026c710498
#> 105                                                                                                RHS:summarise(avgmpg = mean(mpg), avgwt = mean(wt))
#> 106                                                                                                                                         format:rda
#> 107                                                                                                      session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 108                                                                                                                                         format:txt
#> 109                                                                                                                                         format:rda
#> 110                             name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt))
#> 111                                                                                                                                   class:grouped_df
#> 112                                                                                                                                       class:tbl_df
#> 113                                                                                                                                          class:tbl
#> 114                                                                                                                                   class:data.frame
#> 115                                                                                                                                        varname:cyl
#> 116                                                                                                                                         varname:am
#> 117                                                                                                                                     varname:avgmpg
#> 118                                                                                                                                      varname:avgwt
#> 119                                                                                                                           date:2017-03-15 16:40:31
#> 120                                                                                                                                         format:rda
#> 121                                                                                                      session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 122                                                                                                                                         format:txt
#> 123                                                                                                                                         format:rda
#> 124                                                                                                                                       name:res_val
#> 125                                                                                                                                   class:grouped_df
#> 126                                                                                                                                       class:tbl_df
#> 127                                                                                                                                          class:tbl
#> 128                                                                                                                                   class:data.frame
#> 129                                                                                                                                        varname:cyl
#> 130                                                                                                                                         varname:am
#> 131                                                                                                                                     varname:avgmpg
#> 132                                                                                                                                      varname:avgwt
#> 133                                                                                                                           date:2017-03-15 16:40:31
#> 134                                                                                                               LHS:40554c7e8cff1adbacf3405624fd58ce
#> 135                                                                                                                            RHS:filter(avgmpg > 20)
#> 136                                                                                                                                         format:rda
#> 137                                                                                                      session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 138                                                                                                                                         format:txt
#> 139                                                                                                                                         format:rda
#> 140                                                                                                                                         name:__LHS
#> 141                                                                                                                                   class:grouped_df
#> 142                                                                                                                                       class:tbl_df
#> 143                                                                                                                                          class:tbl
#> 144                                                                                                                                   class:data.frame
#> 145                                                                                                                                        varname:cyl
#> 146                                                                                                                                         varname:am
#> 147                                                                                                                                     varname:avgmpg
#> 148                                                                                                                                      varname:avgwt
#> 149                                                                                                                           date:2017-03-15 16:40:31
#> 150                                                                                                                                         format:rda
#> 151                                                                                                      session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 152                                                                                                                                         format:txt
#> 153                                                                                                                                         format:rda
#> 154 name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a% filter(avgmpg >     20)
#> 155                                                                                                                                   class:grouped_df
#> 156                                                                                                                                       class:tbl_df
#> 157                                                                                                                                          class:tbl
#> 158                                                                                                                                   class:data.frame
#> 159                                                                                                                                        varname:cyl
#> 160                                                                                                                                         varname:am
#> 161                                                                                                                                     varname:avgmpg
#> 162                                                                                                                                      varname:avgwt
#> 163                                                                                                                           date:2017-03-15 16:40:31
#> 164                                                                                                                                         format:rda
#> 165                                                                                                      session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#> 166                                                                                                                                         format:txt
#> 167                                                                                                                                         format:rda
#> 168                                                                                                                                       name:res_val
#> 169                                                                                                                                    class:character
#> 170                                                                                                                           date:2017-03-15 16:40:31
#> 171                                                                                                               LHS:177cbfc0062d297ae58cd38566014ba8
#> 172                                                                                                                     RHS:saveToRepo(exampleRepoDir)
#> 173                                                                                                                                         format:rda
#> 174                                                                                                      session_info:2cb4e61dcc8f3c8bb347807aa59461ca
#>             createdDate
#> 1   2017-03-15 16:40:31
#> 2   2017-03-15 16:40:31
#> 3   2017-03-15 16:40:31
#> 4   2017-03-15 16:40:31
#> 5   2017-03-15 16:40:31
#> 6   2017-03-15 16:40:31
#> 7   2017-03-15 16:40:31
#> 8   2017-03-15 16:40:31
#> 9   2017-03-15 16:40:31
#> 10  2017-03-15 16:40:31
#> 11  2017-03-15 16:40:31
#> 12  2017-03-15 16:40:31
#> 13  2017-03-15 16:40:31
#> 14  2017-03-15 16:40:31
#> 15  2017-03-15 16:40:31
#> 16  2017-03-15 16:40:31
#> 17  2017-03-15 16:40:31
#> 18  2017-03-15 16:40:31
#> 19  2017-03-15 16:40:31
#> 20  2017-03-15 16:40:31
#> 21  2017-03-15 16:40:31
#> 22  2017-03-15 16:40:31
#> 23  2017-03-15 16:40:31
#> 24  2017-03-15 16:40:31
#> 25  2017-03-15 16:40:31
#> 26  2017-03-15 16:40:31
#> 27  2017-03-15 16:40:31
#> 28  2017-03-15 16:40:31
#> 29  2017-03-15 16:40:31
#> 30  2017-03-15 16:40:31
#> 31  2017-03-15 16:40:31
#> 32  2017-03-15 16:40:31
#> 33  2017-03-15 16:40:31
#> 34  2017-03-15 16:40:31
#> 35  2017-03-15 16:40:31
#> 36  2017-03-15 16:40:31
#> 37  2017-03-15 16:40:31
#> 38  2017-03-15 16:40:31
#> 39  2017-03-15 16:40:31
#> 40  2017-03-15 16:40:31
#> 41  2017-03-15 16:40:31
#> 42  2017-03-15 16:40:31
#> 43  2017-03-15 16:40:31
#> 44  2017-03-15 16:40:31
#> 45  2017-03-15 16:40:31
#> 46  2017-03-15 16:40:31
#> 47  2017-03-15 16:40:31
#> 48  2017-03-15 16:40:31
#> 49  2017-03-15 16:40:31
#> 50  2017-03-15 16:40:31
#> 51  2017-03-15 16:40:31
#> 52  2017-03-15 16:40:31
#> 53  2017-03-15 16:40:31
#> 54  2017-03-15 16:40:31
#> 55  2017-03-15 16:40:31
#> 56  2017-03-15 16:40:31
#> 57  2017-03-15 16:40:31
#> 58  2017-03-15 16:40:31
#> 59  2017-03-15 16:40:31
#> 60  2017-03-15 16:40:31
#> 61  2017-03-15 16:40:31
#> 62  2017-03-15 16:40:31
#> 63  2017-03-15 16:40:31
#> 64  2017-03-15 16:40:31
#> 65  2017-03-15 16:40:31
#> 66  2017-03-15 16:40:31
#> 67  2017-03-15 16:40:31
#> 68  2017-03-15 16:40:31
#> 69  2017-03-15 16:40:31
#> 70  2017-03-15 16:40:31
#> 71  2017-03-15 16:40:31
#> 72  2017-03-15 16:40:31
#> 73  2017-03-15 16:40:31
#> 74  2017-03-15 16:40:31
#> 75  2017-03-15 16:40:31
#> 76  2017-03-15 16:40:31
#> 77  2017-03-15 16:40:31
#> 78  2017-03-15 16:40:31
#> 79  2017-03-15 16:40:31
#> 80  2017-03-15 16:40:31
#> 81  2017-03-15 16:40:31
#> 82  2017-03-15 16:40:31
#> 83  2017-03-15 16:40:31
#> 84  2017-03-15 16:40:31
#> 85  2017-03-15 16:40:31
#> 86  2017-03-15 16:40:31
#> 87  2017-03-15 16:40:31
#> 88  2017-03-15 16:40:31
#> 89  2017-03-15 16:40:31
#> 90  2017-03-15 16:40:31
#> 91  2017-03-15 16:40:31
#> 92  2017-03-15 16:40:31
#> 93  2017-03-15 16:40:31
#> 94  2017-03-15 16:40:31
#> 95  2017-03-15 16:40:31
#> 96  2017-03-15 16:40:31
#> 97  2017-03-15 16:40:31
#> 98  2017-03-15 16:40:31
#> 99  2017-03-15 16:40:31
#> 100 2017-03-15 16:40:31
#> 101 2017-03-15 16:40:31
#> 102 2017-03-15 16:40:31
#> 103 2017-03-15 16:40:31
#> 104 2017-03-15 16:40:31
#> 105 2017-03-15 16:40:31
#> 106 2017-03-15 16:40:31
#> 107 2017-03-15 16:40:31
#> 108 2017-03-15 16:40:31
#> 109 2017-03-15 16:40:31
#> 110 2017-03-15 16:40:31
#> 111 2017-03-15 16:40:31
#> 112 2017-03-15 16:40:31
#> 113 2017-03-15 16:40:31
#> 114 2017-03-15 16:40:31
#> 115 2017-03-15 16:40:31
#> 116 2017-03-15 16:40:31
#> 117 2017-03-15 16:40:31
#> 118 2017-03-15 16:40:31
#> 119 2017-03-15 16:40:31
#> 120 2017-03-15 16:40:31
#> 121 2017-03-15 16:40:31
#> 122 2017-03-15 16:40:31
#> 123 2017-03-15 16:40:31
#> 124 2017-03-15 16:40:31
#> 125 2017-03-15 16:40:31
#> 126 2017-03-15 16:40:31
#> 127 2017-03-15 16:40:31
#> 128 2017-03-15 16:40:31
#> 129 2017-03-15 16:40:31
#> 130 2017-03-15 16:40:31
#> 131 2017-03-15 16:40:31
#> 132 2017-03-15 16:40:31
#> 133 2017-03-15 16:40:31
#> 134 2017-03-15 16:40:31
#> 135 2017-03-15 16:40:31
#> 136 2017-03-15 16:40:31
#> 137 2017-03-15 16:40:31
#> 138 2017-03-15 16:40:31
#> 139 2017-03-15 16:40:31
#> 140 2017-03-15 16:40:31
#> 141 2017-03-15 16:40:31
#> 142 2017-03-15 16:40:31
#> 143 2017-03-15 16:40:31
#> 144 2017-03-15 16:40:31
#> 145 2017-03-15 16:40:31
#> 146 2017-03-15 16:40:31
#> 147 2017-03-15 16:40:31
#> 148 2017-03-15 16:40:31
#> 149 2017-03-15 16:40:31
#> 150 2017-03-15 16:40:31
#> 151 2017-03-15 16:40:31
#> 152 2017-03-15 16:40:31
#> 153 2017-03-15 16:40:31
#> 154 2017-03-15 16:40:31
#> 155 2017-03-15 16:40:31
#> 156 2017-03-15 16:40:31
#> 157 2017-03-15 16:40:31
#> 158 2017-03-15 16:40:31
#> 159 2017-03-15 16:40:31
#> 160 2017-03-15 16:40:31
#> 161 2017-03-15 16:40:31
#> 162 2017-03-15 16:40:31
#> 163 2017-03-15 16:40:31
#> 164 2017-03-15 16:40:31
#> 165 2017-03-15 16:40:31
#> 166 2017-03-15 16:40:31
#> 167 2017-03-15 16:40:31
#> 168 2017-03-15 16:40:31
#> 169 2017-03-15 16:40:31
#> 170 2017-03-15 16:40:31
#> 171 2017-03-15 16:40:31
#> 172 2017-03-15 16:40:31
#> 173 2017-03-15 16:40:31
#> 174 2017-03-15 16:40:31

# We search for a Tag with default "name" regular expression corresponding to 
# hash md5hash.
getTagsLocal( md5hash = hash, exampleRepoDir )

#> [1] "name:res_val"                                                                                                                                      
#> [2] "name:__LHS"                                                                                                                                        
#> [3] "name:mtcars %a% group_by(cyl, am) %a% select(mpg, cyl, wt, am) %a%     summarise(avgmpg = mean(mpg), avgwt = mean(wt)) %a% filter(avgmpg >     20)"

# Deleting example respository
deleteLocalRepo( exampleRepoDir, TRUE)
rm( exampleRepoDir )

## EXAMPLE with data iris
exampleRepoDir <- tempfile()
createLocalRepo( exampleRepoDir )

#> Directory /tmp/RtmpFa89Nv/file31a7ed6c876 did not exist. Forced to create a new directory.

data(iris)
saveToRepo(iris, repoDir = exampleRepoDir )

#> [1] "ff575c261c949d073b2895b05d1097c3"

showLocalRepo(exampleRepoDir)

#>                            md5hash                             name
#> 1 ff575c261c949d073b2895b05d1097c3                             iris
#> 2 2cb4e61dcc8f3c8bb347807aa59461ca 2cb4e61dcc8f3c8bb347807aa59461ca
#>           createdDate
#> 1 2017-03-15 16:40:31
#> 2 2017-03-15 16:40:32

showLocalRepo(exampleRepoDir, method = "tags")

#>                            artifact
#> 1  ff575c261c949d073b2895b05d1097c3
#> 2  ff575c261c949d073b2895b05d1097c3
#> 3  ff575c261c949d073b2895b05d1097c3
#> 4  ff575c261c949d073b2895b05d1097c3
#> 5  ff575c261c949d073b2895b05d1097c3
#> 6  ff575c261c949d073b2895b05d1097c3
#> 7  ff575c261c949d073b2895b05d1097c3
#> 8  ff575c261c949d073b2895b05d1097c3
#> 9  ff575c261c949d073b2895b05d1097c3
#> 10 2cb4e61dcc8f3c8bb347807aa59461ca
#> 11 ff575c261c949d073b2895b05d1097c3
#> 12 ff575c261c949d073b2895b05d1097c3
#>                                              tag         createdDate
#> 1                                     format:rda 2017-03-15 16:40:31
#> 2                                      name:iris 2017-03-15 16:40:31
#> 3                               class:data.frame 2017-03-15 16:40:31
#> 4                           varname:Sepal.Length 2017-03-15 16:40:31
#> 5                            varname:Sepal.Width 2017-03-15 16:40:31
#> 6                           varname:Petal.Length 2017-03-15 16:40:31
#> 7                            varname:Petal.Width 2017-03-15 16:40:31
#> 8                                varname:Species 2017-03-15 16:40:31
#> 9                       date:2017-03-15 16:40:31 2017-03-15 16:40:31
#> 10                                    format:rda 2017-03-15 16:40:32
#> 11 session_info:2cb4e61dcc8f3c8bb347807aa59461ca 2017-03-15 16:40:32
#> 12                                    format:txt 2017-03-15 16:40:32

# We can notice that there is only one md5hash 
# (and second for archiveSessionInfo) in repo so we will use it
hash <- showLocalRepo(exampleRepoDir)[1,1]

# We search for a Tag with "varname" regular expression corresponding to 
# hash md5hash.
getTagsLocal( md5hash = hash, exampleRepoDir, tag = "varname" )

#> [1] "varname:Sepal.Length" "varname:Sepal.Width"  "varname:Petal.Length"
#> [4] "varname:Petal.Width"  "varname:Species"     

# There are 5 different Tags with "varname" regular expression

# We needn't use the whole expression "varname". We may use its abbreviation
# and get the same result.
getTagsLocal( md5hash = hash, exampleRepoDir, tag = "varna" )

#> [1] "varname:Sepal.Length" "varname:Sepal.Width"  "varname:Petal.Length"
#> [4] "varname:Petal.Width"  "varname:Species"     

deleteLocalRepo( exampleRepoDir, TRUE)
rm( exampleRepoDir )

### Remote version
## EXAMPLE: pbiecek archivist repository on GitHub

showRemoteRepo(user="pbiecek", repo="archivist")

#>                             md5hash                             name
#> 1  ff575c261c949d073b2895b05d1097c3                             iris
#> 2  f7804cb74efa70d9795ace8c32e333c3                        myplot123
#> 3  5c3b6d82ab29e7b3a2137b6a537e55ec 5c3b6d82ab29e7b3a2137b6a537e55ec
#> 4  cd6557c6163a6f9800f308f343e75e72                            model
#> 5  7a761a2ae54f3d90060a9f6ca04b3506 7a761a2ae54f3d90060a9f6ca04b3506
#> 6  db864b3dcc09888f5f7f2750b4ab553d                             agn1
#> 7  c2380a7b8c090d6564ff12cdd60aaf7f c2380a7b8c090d6564ff12cdd60aaf7f
#> 8  c1ce897de77e8cb793a69b54e7f2b8b5                           fannyx
#> 9  ebe92bb4f9af6c343da4986f84c42b9b ebe92bb4f9af6c343da4986f84c42b9b
#> 10 2ae5f73c0b979532555fb3800458a4d4                             lda1
#> 11 53a9f3aa4235e111524dda17aad2ee3a 53a9f3aa4235e111524dda17aad2ee3a
#> 12 fdf1922d88bb7369896886dfa7da16f6                             qda1
#> 13 4d504b6c315c9be389f2e84186fa64c5 4d504b6c315c9be389f2e84186fa64c5
#> 14 1e05d1a3f103da0d99fba79643f71487                          glmnet1
#> 15 942ca3ecb05524e95375728c114bf619 942ca3ecb05524e95375728c114bf619
#> 16 134ecbbe2a8814d98f0c2758000c408e                     trellis.plot
#> 17 f49104cd336e6417f89141855c0fe4a7 f49104cd336e6417f89141855c0fe4a7
#> 18 df598bf17a3f149e004eede9c68313fa                        this.test
#> 19 3c6ee510336926b494cf8375b903246a 3c6ee510336926b494cf8375b903246a
#> 20 1d85e04b24e5129f0b7577b270d24ab4                            myFit
#> 21 917b56fd9edd81e088f8f7ada48cda95 917b56fd9edd81e088f8f7ada48cda95
#> 22 617ded4953ac986524a1c24703363980 617ded4953ac986524a1c24703363980
#> 23 5629bc43e36d219b613076b17c665eda 5629bc43e36d219b613076b17c665eda
#>            createdDate
#> 1  2015-11-16 12:47:13
#> 2  2015-11-16 12:47:14
#> 3  2015-11-16 12:47:14
#> 4  2015-11-16 12:47:14
#> 5  2015-11-16 12:47:14
#> 6  2015-11-16 12:47:14
#> 7  2015-11-16 12:47:14
#> 8  2015-11-16 12:47:14
#> 9  2015-11-16 12:47:14
#> 10 2015-11-16 12:47:15
#> 11 2015-11-16 12:47:15
#> 12 2015-11-16 12:47:15
#> 13 2015-11-16 12:47:15
#> 14 2015-11-16 12:47:16
#> 15 2015-11-16 12:47:16
#> 16 2015-11-16 12:47:18
#> 17 2015-11-16 12:47:18
#> 18 2015-11-16 12:47:19
#> 19 2015-11-16 12:47:19
#> 20 2015-11-16 12:47:19
#> 21 2015-11-16 12:47:19
#> 22 2015-11-16 12:47:24
#> 23 2015-11-16 12:47:24

# We search for a Tag with default "name" regular expression corresponding to 
# "cd6557c6163a6f9800f308f343e75e72" md5hash.
getTagsRemote( "cd6557c6163a6f9800f308f343e75e72",
                user="pbiecek", repo="archivist")

#> [1] "name:model"

## EXAMPLE: many archivist-like Repositories on one Github repository
# We search for a Tag with default "name" regular expression corresponding to 
# "ff575c261c949d073b2895b05d1097c3" md5hash.
getTagsRemote("ff575c261c949d073b2895b05d1097c3", user="MarcinKosinski",
               repo="Museum", branch="master", subdir="ex1")

#> [1] "name:iris"