Save an Artifact into a Repository

saveToLocalRepo function saves desired artifacts to the local Repository in a given directory. To learn more about artifacts visit archivist-package.

saveToLocalRepo(artifact, repoDir = aoptions("repoDir"), archiveData = TRUE,
  archiveTags = TRUE, archiveMiniature = TRUE, archiveSessionInfo = TRUE,
  force = TRUE, value = FALSE, ..., userTags = c(),
  silent = aoptions("silent"), ascii = FALSE,
  artifactName = deparse(substitute(artifact)))

saveToRepo(artifact, repoDir = aoptions("repoDir"), archiveData = TRUE,
  archiveTags = TRUE, archiveMiniature = TRUE, archiveSessionInfo = TRUE,
  force = TRUE, value = FALSE, ..., userTags = c(),
  silent = aoptions("silent"), ascii = FALSE,
  artifactName = deparse(substitute(artifact)))

asave(artifact, repoDir = aoptions("repoDir"), archiveData = TRUE,
  archiveTags = TRUE, archiveMiniature = TRUE, archiveSessionInfo = TRUE,
  force = TRUE, value = FALSE, ..., userTags = c(),
  silent = aoptions("silent"), ascii = FALSE,
  artifactName = deparse(substitute(artifact)))

Arguments

artifact
An arbitrary R artifact to be saved. For supported artifacts see details.

repoDir
A character denoting an existing directory in which an artifact will be saved.

archiveData
A logical value denoting whether to archive the data from the artifact.

archiveTags
A logical value denoting whether to archive Tags from the artifact.

archiveMiniature
A logical value denoting whether to archive a miniature of the artifact.

archiveSessionInfo
A logical value denoting whether to archive the session info that describes the context in this given artifact was created.

force
A logical value denoting whether to archive artifact if it was already archived in a Repository.

value
A logical value. Should the result be (default value = FALSE) the md5hash of a stored artifact or should the result be an input artifact (value = TRUE), so that valueing code can be used. See examples.

...
Graphical parameters denoting width and height of a miniature. See details. Further arguments passed to head. See Details section about firtsRows parameter

userTags
A character vector with Tags. These Tags will be added to the repository along with the artifact.

silent
If TRUE produces no warnings.

ascii
A logical value. An ascii argument is passed to save function.

artifactName
The name of the artifact with which it should be archived. If NULL then object's MD5 hash will be used instead.

Value

As a result of calling this function a character string is returned, which determines the md5hash of the artifact. If archiveData is TRUE, the result will also have an attribute, named data, which determines md5hash of the data needed to compute the artifact.

Details

saveToLocalRepo function saves desired artifacts to the local Repository in a given directory. Artifacts are saved in the local Repository, which is a SQLite database named backpack. After every saveToLocalRepo call the database is refreshed, so the artifact is available immediately in the database for other collaborators. Each artifact is archived in a md5hash.rda file. This file will be saved in a folder (under repoDir directory) named gallery. For each artifact, md5hash is a unique string of length 32 that is produced by digest function, which uses a cryptographical MD5 hash algorithm.

By default, a miniature of an artifact and (if possible) a data set needed to compute this artifact are extracted. They are also going to be saved in a file named by their md5hash in the gallery folder that exists in the directory specified in the repoDir argument. Moreover, a specific Tag-relation is going to be added to the backpack dataset in case there is a need to load the artifact with it's related data set - see loadFromLocalRepo or loadFromRemoteRepo. Default settings may be changed by using the archiveData, archiveTag or archiveMiniature arguments with the FALSE value. Tags are artifact's attributes, different for various artifact's classes. For more detailed information check Tags

Archived artifact can be searched in the backpack dataset by using the searchInLocalRepo or searchInRemoteRepo functions. Artifacts can be searched by their Tags, names, classes or archiving date. firstRows parameter.

If the artifact is of class data.frame or user set archiveData = TRUE for artifact that stores data within it, it is possible to specify how many rows of that data (or that data.frame) should be archived in a miniature. This can be done by adding the argument firstRows with the n corresponding to the number of rows (as in head). Note that, the data can be extracted only from the artifacts that are supported by the archivist package; see Tags.

Graphical parameters.

If the artifact is of class lattice or ggplot, and archiveMiniature = TRUE, then it is possible to set the miniature's width and height parameters. By default they are set to width = 800, height = 600.

Supported artifact's classes are listed here Tags.

Note

In the following way one can specify his own Tags for artifacts by setting artifact's attribute before call of the saveToLocalRepo function: attr(x, "tags" ) = c( "name1", "name2" ), where x is an artifact and name1, name2 are Tags specified by a user. It can be also done in a new, simpler way by using userTags parameter like this:

saveToLocalRepo(model, repoDir, userTags = c("my_model", "do not delete")).

Specifing additional Tags by attributes can be beneficial when one uses addHooksToPrint.

Important: if one wants to archive data from artifacts which is one of: survfit, glmnet, qda, lda, trellis, htest class, and this dataset is transformed within the artifact's formula then saveToLocalRepo will not archive this dataset. saveToLocalRepo only archives datasets that already exist in any of R environments.

Example: The data set will not be archived here.

z <- lda(Sp ~ ., Iris, prior = c(1,1,1)/3, subset = train[,-8])

saveToLocalRepo( z, repoDir )

Example: The data set will be archived here.

train2 <- train[,-8]

z <- lda(Sp ~ ., Iris, prior = c(1,1,1)/3, subset = train2)

saveToLocalRepo( z, repoDir )

Contact

Bug reports and feature requests can be sent to https://github.com/pbiecek/archivist/issues

References

Biecek P and Kosinski M (2017). "archivist: An R Package for Managing, Recording and Restoring Data Analysis Results." _Journal of Statistical Software_, *82*(11), pp. 1-28. doi: 10.18637/jss.v082.i11 (URL: http://doi.org/10.18637/jss.v082.i11). URL https://github.com/pbiecek/archivist

See also

For more detailed information check the archivist package Use Cases. The list of supported artifacts and their tags is available on wiki on archivist Github Repository.

Other archivist: Repository, Tags, %a%, addHooksToPrint, addTagsRepo, aformat, ahistory, alink, aoptions, archivist-package, areadLocal, aread, asearchLocal, asearch, asession, atrace, cache, copyLocalRepo, createLocalRepo, createMDGallery, deleteLocalRepo, getRemoteHook, getTagsLocal, loadFromLocalRepo, md5hash, removeTagsRepo, restoreLibs, rmFromLocalRepo, searchInLocalRepo, setLocalRepo, shinySearchInLocalRepo, showLocalRepo, splitTagsLocal, summaryLocalRepo, zipLocalRepo

Examples

exampleRepoDir <- tempfile()
createLocalRepo(repoDir = exampleRepoDir)
data(swiss)
saveToLocalRepo(swiss, repoDir=exampleRepoDir, archiveSessionInfo = TRUE)

#> [1] "4c43fa8a4d8f0cbf65353e397f37338c"

showLocalRepo(method = "md5hashes", repoDir = exampleRepoDir)

#>                            md5hash                             name
#> 1 4c43fa8a4d8f0cbf65353e397f37338c                            swiss
#> 2 7e3aefa426aa39c40a63405b6dfec950 7e3aefa426aa39c40a63405b6dfec950
#>           createdDate
#> 1 2017-11-25 01:47:21
#> 2 2017-11-25 01:47:21

showLocalRepo(method = "tags", repoDir = exampleRepoDir)

#>                            artifact
#> 1  4c43fa8a4d8f0cbf65353e397f37338c
#> 2  4c43fa8a4d8f0cbf65353e397f37338c
#> 3  4c43fa8a4d8f0cbf65353e397f37338c
#> 4  4c43fa8a4d8f0cbf65353e397f37338c
#> 5  4c43fa8a4d8f0cbf65353e397f37338c
#> 6  4c43fa8a4d8f0cbf65353e397f37338c
#> 7  4c43fa8a4d8f0cbf65353e397f37338c
#> 8  4c43fa8a4d8f0cbf65353e397f37338c
#> 9  4c43fa8a4d8f0cbf65353e397f37338c
#> 10 4c43fa8a4d8f0cbf65353e397f37338c
#> 11 7e3aefa426aa39c40a63405b6dfec950
#> 12 4c43fa8a4d8f0cbf65353e397f37338c
#> 13 4c43fa8a4d8f0cbf65353e397f37338c
#>                                              tag         createdDate
#> 1                                     format:rda 2017-11-25 01:47:21
#> 2                                     name:swiss 2017-11-25 01:47:21
#> 3                               class:data.frame 2017-11-25 01:47:21
#> 4                              varname:Fertility 2017-11-25 01:47:21
#> 5                            varname:Agriculture 2017-11-25 01:47:21
#> 6                            varname:Examination 2017-11-25 01:47:21
#> 7                              varname:Education 2017-11-25 01:47:21
#> 8                               varname:Catholic 2017-11-25 01:47:21
#> 9                       varname:Infant.Mortality 2017-11-25 01:47:21
#> 10                      date:2017-11-25 01:47:21 2017-11-25 01:47:21
#> 11                                    format:rda 2017-11-25 01:47:21
#> 12 session_info:7e3aefa426aa39c40a63405b6dfec950 2017-11-25 01:47:21
#> 13                                    format:txt 2017-11-25 01:47:21

loadFromLocalRepo(md5hash = '2a6e492cb6982f230e48cf46023e2e4f',
  repoDir = system.file("graphGallery", package = "archivist"), value = TRUE) -> model

saveToLocalRepo(model, repoDir=exampleRepoDir,
             userTags = c("do not delete", "my favourite model"))

#> [1] "2a6e492cb6982f230e48cf46023e2e4f"
#> attr(,"data")
#> [1] "afb2550d0f886f0cf3b050f04c5cd4f8"

aoptions('repoDir', system.file("graphGallery", package = "archivist"))

#> [1] "/Users/pbiecek/GitHub/archivist/inst/graphGallery"

showLocalRepo(method = "tags")

#>                             artifact
#> 1   d9313a0de3e2980201a8971e3384ff26
#> 2   ff575c261c949d073b2895b05d1097c3
#> 3   ff575c261c949d073b2895b05d1097c3
#> 4   ff575c261c949d073b2895b05d1097c3
#> 5   2a6e492cb6982f230e48cf46023e2e4f
#> 6   2a6e492cb6982f230e48cf46023e2e4f
#> 7   2a6e492cb6982f230e48cf46023e2e4f
#> 8   2a6e492cb6982f230e48cf46023e2e4f
#> 9   2a6e492cb6982f230e48cf46023e2e4f
#> 10  2a6e492cb6982f230e48cf46023e2e4f
#> 11  2a6e492cb6982f230e48cf46023e2e4f
#> 12  2a6e492cb6982f230e48cf46023e2e4f
#> 13  2a6e492cb6982f230e48cf46023e2e4f
#> 14  2a6e492cb6982f230e48cf46023e2e4f
#> 15  93ecfdf1436932e2860c6dbdf2abc2ad
#> 16  2a6e492cb6982f230e48cf46023e2e4f
#> 17  afb2550d0f886f0cf3b050f04c5cd4f8
#> 18  afb2550d0f886f0cf3b050f04c5cd4f8
#> 19  afb2550d0f886f0cf3b050f04c5cd4f8
#> 20  2a6e492cb6982f230e48cf46023e2e4f
#> 21  18a98048f0584469483afb65294ce3ed
#> 22  18a98048f0584469483afb65294ce3ed
#> 23  18a98048f0584469483afb65294ce3ed
#> 24  18a98048f0584469483afb65294ce3ed
#> 25  18a98048f0584469483afb65294ce3ed
#> 26  18a98048f0584469483afb65294ce3ed
#> 27  18a98048f0584469483afb65294ce3ed
#> 28  18a98048f0584469483afb65294ce3ed
#> 29  2b1fe80a89976f859ee2247723889256
#> 30  18a98048f0584469483afb65294ce3ed
#> 31  9325d4e26da1002bbb0ee75d099e6722
#> 32  b17388f623ba94ef7ca4254ed39c92a7
#> 33  9325d4e26da1002bbb0ee75d099e6722
#> 34  9325d4e26da1002bbb0ee75d099e6722
#> 35  9325d4e26da1002bbb0ee75d099e6722
#> 36  18a98048f0584469483afb65294ce3ed
#> 37  b17388f623ba94ef7ca4254ed39c92a7
#> 38  ff575c261c949d073b2895b05d1097c3
#> 39  b17388f623ba94ef7ca4254ed39c92a7
#> 40  ff575c261c949d073b2895b05d1097c3
#> 41  ff575c261c949d073b2895b05d1097c3
#> 42  ff575c261c949d073b2895b05d1097c3
#> 43  ff575c261c949d073b2895b05d1097c3
#> 44  ff575c261c949d073b2895b05d1097c3
#> 45  ff575c261c949d073b2895b05d1097c3
#> 46  ff575c261c949d073b2895b05d1097c3
#> 47  ff575c261c949d073b2895b05d1097c3
#> 48  ff575c261c949d073b2895b05d1097c3
#> 49  ff575c261c949d073b2895b05d1097c3
#> 50  ff575c261c949d073b2895b05d1097c3
#> 51  ff575c261c949d073b2895b05d1097c3
#> 52  fe1dd03597308686a7fff7161f644b6b
#> 53  ff575c261c949d073b2895b05d1097c3
#> 54  ff575c261c949d073b2895b05d1097c3
#> 55  d3696e13d15223c7d0bbccb33cc20a11
#> 56  d3696e13d15223c7d0bbccb33cc20a11
#> 57  d3696e13d15223c7d0bbccb33cc20a11
#> 58  d3696e13d15223c7d0bbccb33cc20a11
#> 59  d3696e13d15223c7d0bbccb33cc20a11
#> 60  d3696e13d15223c7d0bbccb33cc20a11
#> 61  d3696e13d15223c7d0bbccb33cc20a11
#> 62  d3696e13d15223c7d0bbccb33cc20a11
#> 63  d3696e13d15223c7d0bbccb33cc20a11
#> 64  d3696e13d15223c7d0bbccb33cc20a11
#> 65  fe1dd03597308686a7fff7161f644b6b
#> 66  d3696e13d15223c7d0bbccb33cc20a11
#> 67  d3696e13d15223c7d0bbccb33cc20a11
#> 68  d3696e13d15223c7d0bbccb33cc20a11
#> 69  d3696e13d15223c7d0bbccb33cc20a11
#> 70  d3696e13d15223c7d0bbccb33cc20a11
#> 71  d3696e13d15223c7d0bbccb33cc20a11
#> 72  d3696e13d15223c7d0bbccb33cc20a11
#> 73  d3696e13d15223c7d0bbccb33cc20a11
#> 74  d3696e13d15223c7d0bbccb33cc20a11
#> 75  d3696e13d15223c7d0bbccb33cc20a11
#> 76  d3696e13d15223c7d0bbccb33cc20a11
#> 77  fe1dd03597308686a7fff7161f644b6b
#> 78  d3696e13d15223c7d0bbccb33cc20a11
#> 79  d3696e13d15223c7d0bbccb33cc20a11
#> 80  990861c7c27812ee959f10e5f76fe2c3
#> 81  990861c7c27812ee959f10e5f76fe2c3
#> 82  990861c7c27812ee959f10e5f76fe2c3
#> 83  990861c7c27812ee959f10e5f76fe2c3
#> 84  990861c7c27812ee959f10e5f76fe2c3
#> 85  990861c7c27812ee959f10e5f76fe2c3
#> 86  990861c7c27812ee959f10e5f76fe2c3
#> 87  990861c7c27812ee959f10e5f76fe2c3
#> 88  990861c7c27812ee959f10e5f76fe2c3
#> 89  990861c7c27812ee959f10e5f76fe2c3
#> 90  fe1dd03597308686a7fff7161f644b6b
#> 91  990861c7c27812ee959f10e5f76fe2c3
#> 92  990861c7c27812ee959f10e5f76fe2c3
#> 93  990861c7c27812ee959f10e5f76fe2c3
#> 94  990861c7c27812ee959f10e5f76fe2c3
#> 95  990861c7c27812ee959f10e5f76fe2c3
#> 96  990861c7c27812ee959f10e5f76fe2c3
#> 97  990861c7c27812ee959f10e5f76fe2c3
#> 98  990861c7c27812ee959f10e5f76fe2c3
#> 99  990861c7c27812ee959f10e5f76fe2c3
#> 100 990861c7c27812ee959f10e5f76fe2c3
#> 101 990861c7c27812ee959f10e5f76fe2c3
#> 102 fe1dd03597308686a7fff7161f644b6b
#> 103 990861c7c27812ee959f10e5f76fe2c3
#> 104 990861c7c27812ee959f10e5f76fe2c3
#> 105 050e41ec3bc40b3004bc6bdd356acae7
#> 106 050e41ec3bc40b3004bc6bdd356acae7
#> 107 050e41ec3bc40b3004bc6bdd356acae7
#> 108 050e41ec3bc40b3004bc6bdd356acae7
#> 109 050e41ec3bc40b3004bc6bdd356acae7
#> 110 050e41ec3bc40b3004bc6bdd356acae7
#> 111 050e41ec3bc40b3004bc6bdd356acae7
#> 112 050e41ec3bc40b3004bc6bdd356acae7
#> 113 050e41ec3bc40b3004bc6bdd356acae7
#> 114 050e41ec3bc40b3004bc6bdd356acae7
#> 115 050e41ec3bc40b3004bc6bdd356acae7
#> 116 050e41ec3bc40b3004bc6bdd356acae7
#> 117 050e41ec3bc40b3004bc6bdd356acae7
#> 118 050e41ec3bc40b3004bc6bdd356acae7
#> 119 fe1dd03597308686a7fff7161f644b6b
#> 120 050e41ec3bc40b3004bc6bdd356acae7
#> 121 2362b67ee86693d7326c95d3526a5cd8
#> 122 ff575c261c949d073b2895b05d1097c3
#> 123 ff575c261c949d073b2895b05d1097c3
#> 124 ff575c261c949d073b2895b05d1097c3
#> 125 1862bb3adaf78b7b3b9a30b46526bf7e
#> 126 ff575c261c949d073b2895b05d1097c3
#> 127 ff575c261c949d073b2895b05d1097c3
#> 128 ff575c261c949d073b2895b05d1097c3
#> 129 1862bb3adaf78b7b3b9a30b46526bf7e
#> 130 ff575c261c949d073b2895b05d1097c3
#> 131 ff575c261c949d073b2895b05d1097c3
#> 132 ff575c261c949d073b2895b05d1097c3
#> 133 7f3453331910e3f321ef97d87adb5bad
#> 134 7f3453331910e3f321ef97d87adb5bad
#> 135 7f3453331910e3f321ef97d87adb5bad
#> 136 7f3453331910e3f321ef97d87adb5bad
#> 137 7f3453331910e3f321ef97d87adb5bad
#> 138 7f3453331910e3f321ef97d87adb5bad
#> 139 7f3453331910e3f321ef97d87adb5bad
#> 140 78ce2c8ede25776f1f22c538258eed6d
#> 141 7f3453331910e3f321ef97d87adb5bad
#> 142 ff575c261c949d073b2895b05d1097c3
#> 143 ff575c261c949d073b2895b05d1097c3
#> 144 ff575c261c949d073b2895b05d1097c3
#> 145 7f3453331910e3f321ef97d87adb5bad
#> 146 369227e67f9164dcbe934dadf2b53cc2
#> 147 369227e67f9164dcbe934dadf2b53cc2
#> 148 369227e67f9164dcbe934dadf2b53cc2
#> 149 369227e67f9164dcbe934dadf2b53cc2
#> 150 369227e67f9164dcbe934dadf2b53cc2
#> 151 369227e67f9164dcbe934dadf2b53cc2
#> 152 369227e67f9164dcbe934dadf2b53cc2
#> 153 78ce2c8ede25776f1f22c538258eed6d
#> 154 369227e67f9164dcbe934dadf2b53cc2
#> 155 ff575c261c949d073b2895b05d1097c3
#> 156 ff575c261c949d073b2895b05d1097c3
#> 157 ff575c261c949d073b2895b05d1097c3
#> 158 369227e67f9164dcbe934dadf2b53cc2
#> 159 ff575c261c949d073b2895b05d1097c3
#> 160 ff575c261c949d073b2895b05d1097c3
#> 161 ff575c261c949d073b2895b05d1097c3
#> 162 ff575c261c949d073b2895b05d1097c3
#> 163 ff575c261c949d073b2895b05d1097c3
#> 164 ff575c261c949d073b2895b05d1097c3
#> 165 ff575c261c949d073b2895b05d1097c3
#> 166 ff575c261c949d073b2895b05d1097c3
#> 167 ff575c261c949d073b2895b05d1097c3
#> 168 c648ca6d0bb605062f5b9416818262e7
#> 169 ff575c261c949d073b2895b05d1097c3
#> 170 ff575c261c949d073b2895b05d1097c3
#> 171 f2152da1fea50f1e76643d0569b3ed84
#> 172 f2152da1fea50f1e76643d0569b3ed84
#> 173 f2152da1fea50f1e76643d0569b3ed84
#> 174 f2152da1fea50f1e76643d0569b3ed84
#> 175 f2152da1fea50f1e76643d0569b3ed84
#> 176 f2152da1fea50f1e76643d0569b3ed84
#> 177 f2152da1fea50f1e76643d0569b3ed84
#> 178 f2152da1fea50f1e76643d0569b3ed84
#> 179 7e3aefa426aa39c40a63405b6dfec950
#> 180 f2152da1fea50f1e76643d0569b3ed84
#> 181 5bff91d3bb65f86cc334bc504c5ccb84
#> 182 5bff91d3bb65f86cc334bc504c5ccb84
#> 183 5bff91d3bb65f86cc334bc504c5ccb84
#> 184 f2152da1fea50f1e76643d0569b3ed84
#>                                                                                     tag
#> 1                                                                            format:rda
#> 2                                                                            format:rda
#> 3                                                                            format:txt
#> 4                                         relationWith:600bda83cb840947976bd1ce3a11879d
#> 5                                                                            format:rda
#> 6                                                                            name:model
#> 7                                                                              class:lm
#> 8                                                                  coefname:(Intercept)
#> 9                                                                 coefname:Sepal.Length
#> 10                                                           coefname:Speciesversicolor
#> 11                                                            coefname:Speciesvirginica
#> 12                                                                               rank:4
#> 13                                                                      df.residual:146
#> 14                                                             date:2016-02-07 01:00:33
#> 15                                                                           format:rda
#> 16                                        session_info:93ecfdf1436932e2860c6dbdf2abc2ad
#> 17                                                                           format:rda
#> 18                                                                           format:txt
#> 19                                        relationWith:2a6e492cb6982f230e48cf46023e2e4f
#> 20                                                                           format:txt
#> 21                                                                           format:rda
#> 22                                                                          name:model2
#> 23                                                                             class:lm
#> 24                                                                 coefname:(Intercept)
#> 25                                                                coefname:Sepal.Length
#> 26                                                                               rank:2
#> 27                                                                      df.residual:148
#> 28                                                             date:2016-02-08 15:22:15
#> 29                                                                           format:rda
#> 30                                        session_info:2b1fe80a89976f859ee2247723889256
#> 31                                                                           format:rda
#> 32                                                                           format:rda
#> 33                                        session_info:b17388f623ba94ef7ca4254ed39c92a7
#> 34                                                                           format:txt
#> 35                                        relationWith:18a98048f0584469483afb65294ce3ed
#> 36                                                                           format:txt
#> 37                                                                           format:rda
#> 38                                                                           format:rda
#> 39                                                                           format:rda
#> 40                                        session_info:b17388f623ba94ef7ca4254ed39c92a7
#> 41                                                                           format:txt
#> 42                                        relationWith:70f4f83dbfef2a8afb2f495208d0a60f
#> 43                                                                           format:rda
#> 44                                                                     class:data.frame
#> 45                                                                 varname:Sepal.Length
#> 46                                                                  varname:Sepal.Width
#> 47                                                                 varname:Petal.Length
#> 48                                                                  varname:Petal.Width
#> 49                                                                      varname:Species
#> 50                                                             date:2016-02-08 15:57:32
#> 51                                                                            name:iris
#> 52                                                                           format:rda
#> 53                                        session_info:fe1dd03597308686a7fff7161f644b6b
#> 54                                                                           format:txt
#> 55                                                                           format:rda
#> 56                                                                     class:data.frame
#> 57                                                                 varname:Sepal.Length
#> 58                                                                  varname:Sepal.Width
#> 59                                                                 varname:Petal.Length
#> 60                                                                  varname:Petal.Width
#> 61                                                                      varname:Species
#> 62                                                             date:2016-02-08 15:57:32
#> 63                                                 LHS:ff575c261c949d073b2895b05d1097c3
#> 64                                                         RHS:filter(Sepal.Length < 6)
#> 65                                                                           format:rda
#> 66                                        session_info:fe1dd03597308686a7fff7161f644b6b
#> 67                                                                           format:txt
#> 68                                                                           format:rda
#> 69                                                                     class:data.frame
#> 70                                                                 varname:Sepal.Length
#> 71                                                                  varname:Sepal.Width
#> 72                                                                 varname:Petal.Length
#> 73                                                                  varname:Petal.Width
#> 74                                                                      varname:Species
#> 75                                                             date:2016-02-08 15:57:32
#> 76                                               name:iris %a% filter(Sepal.Length < 6)
#> 77                                                                           format:rda
#> 78                                        session_info:fe1dd03597308686a7fff7161f644b6b
#> 79                                                                           format:txt
#> 80                                                                           format:rda
#> 81                                                                             class:lm
#> 82                                                                 coefname:(Intercept)
#> 83                                                           coefname:Speciesversicolor
#> 84                                                            coefname:Speciesvirginica
#> 85                                                                               rank:3
#> 86                                                                       df.residual:80
#> 87                                                             date:2016-02-08 15:57:32
#> 88                                                 LHS:d3696e13d15223c7d0bbccb33cc20a11
#> 89                                             RHS:lm(Petal.Length ~ Species, data = .)
#> 90                                                                           format:rda
#> 91                                        session_info:fe1dd03597308686a7fff7161f644b6b
#> 92                                                                           format:txt
#> 93                                                                           format:rda
#> 94                                                                             class:lm
#> 95                                                                 coefname:(Intercept)
#> 96                                                           coefname:Speciesversicolor
#> 97                                                            coefname:Speciesvirginica
#> 98                                                                               rank:3
#> 99                                                                       df.residual:80
#> 100                                                            date:2016-02-08 15:57:32
#> 101 name:iris %a% filter(Sepal.Length < 6) %a% lm(Petal.Length ~ Species,     data = .)
#> 102                                                                          format:rda
#> 103                                       session_info:fe1dd03597308686a7fff7161f644b6b
#> 104                                                                          format:txt
#> 105                                                                          format:rda
#> 106                                                                    class:summary.lm
#> 107                                                                        sigma:0.2803
#> 108                                                                                df:3
#> 109                                                                               df:80
#> 110                                                                                df:3
#> 111                                                                          R^2:0.9604
#> 112                                                                 adjusted R^2:0.9594
#> 113                                                                    fstatistic:970.8
#> 114                                                                     fstatistic.df:2
#> 115                                                                    fstatistic.df:80
#> 116                                                            date:2016-02-08 15:57:32
#> 117                                                LHS:990861c7c27812ee959f10e5f76fe2c3
#> 118                                                                       RHS:summary()
#> 119                                                                          format:rda
#> 120                                       session_info:fe1dd03597308686a7fff7161f644b6b
#> 121                                                                          format:rda
#> 122                                                                          format:rda
#> 123                                                                          format:txt
#> 124                                       relationWith:f05f0ed0662fe01850ec1b928830ef32
#> 125                                                                          format:rda
#> 126                                                                          format:rda
#> 127                                                                          format:txt
#> 128                                       relationWith:13b2724139eb2c62578b4dab0d7b2cea
#> 129                                                                          format:rda
#> 130                                                                          format:rda
#> 131                                                                          format:txt
#> 132                                       relationWith:13b2724139eb2c62578b4dab0d7b2cea
#> 133                                                                          format:rda
#> 134                                                                            name:pl1
#> 135                                                                            class:gg
#> 136                                                                        class:ggplot
#> 137                                                                 labelx:Sepal.Length
#> 138                                                                 labely:Petal.Length
#> 139                                                            date:2016-12-31 15:39:59
#> 140                                                                          format:rda
#> 141                                       session_info:78ce2c8ede25776f1f22c538258eed6d
#> 142                                                                          format:rda
#> 143                                                                          format:txt
#> 144                                       relationWith:7f3453331910e3f321ef97d87adb5bad
#> 145                                                                          format:png
#> 146                                                                          format:rda
#> 147                                                                            name:pl2
#> 148                                                                            class:gg
#> 149                                                                        class:ggplot
#> 150                                                                 labelx:Sepal.Length
#> 151                                                                 labely:Petal.Length
#> 152                                                            date:2016-12-31 15:40:19
#> 153                                                                          format:rda
#> 154                                       session_info:78ce2c8ede25776f1f22c538258eed6d
#> 155                                                                          format:rda
#> 156                                                                          format:txt
#> 157                                       relationWith:369227e67f9164dcbe934dadf2b53cc2
#> 158                                                                          format:png
#> 159                                                                          format:rda
#> 160                                                                           name:iris
#> 161                                                                    class:data.frame
#> 162                                                                varname:Sepal.Length
#> 163                                                                 varname:Sepal.Width
#> 164                                                                varname:Petal.Length
#> 165                                                                 varname:Petal.Width
#> 166                                                                     varname:Species
#> 167                                                            date:2017-11-21 14:20:10
#> 168                                                                          format:rda
#> 169                                       session_info:c648ca6d0bb605062f5b9416818262e7
#> 170                                                                          format:txt
#> 171                                                                          format:rda
#> 172                                                                       name:get("z")
#> 173                                                                            class:lm
#> 174                                                                coefname:(Intercept)
#> 175                                                                coefname:Sepal.Width
#> 176                                                                              rank:2
#> 177                                                                     df.residual:148
#> 178                                                            date:2017-11-25 01:47:10
#> 179                                                                          format:rda
#> 180                                       session_info:7e3aefa426aa39c40a63405b6dfec950
#> 181                                                                          format:rda
#> 182                                                                          format:txt
#> 183                                       relationWith:f2152da1fea50f1e76643d0569b3ed84
#> 184                                                                          format:txt
#>             createdDate
#> 1   2016-02-07 00:48:44
#> 2   2016-02-07 00:48:44
#> 3   2016-02-07 00:48:44
#> 4   2016-02-07 00:48:44
#> 5   2016-02-07 01:00:33
#> 6   2016-02-07 01:00:33
#> 7   2016-02-07 01:00:33
#> 8   2016-02-07 01:00:33
#> 9   2016-02-07 01:00:33
#> 10  2016-02-07 01:00:33
#> 11  2016-02-07 01:00:33
#> 12  2016-02-07 01:00:33
#> 13  2016-02-07 01:00:33
#> 14  2016-02-07 01:00:33
#> 15  2016-02-07 01:00:33
#> 16  2016-02-07 01:00:33
#> 17  2016-02-07 01:00:33
#> 18  2016-02-07 01:00:33
#> 19  2016-02-07 01:00:33
#> 20  2016-02-07 01:00:33
#> 21  2016-02-08 15:22:15
#> 22  2016-02-08 15:22:15
#> 23  2016-02-08 15:22:15
#> 24  2016-02-08 15:22:15
#> 25  2016-02-08 15:22:15
#> 26  2016-02-08 15:22:15
#> 27  2016-02-08 15:22:15
#> 28  2016-02-08 15:22:15
#> 29  2016-02-08 15:22:15
#> 30  2016-02-08 15:22:15
#> 31  2016-02-08 15:22:15
#> 32  2016-02-08 15:22:15
#> 33  2016-02-08 15:22:15
#> 34  2016-02-08 15:22:15
#> 35  2016-02-08 15:22:15
#> 36  2016-02-08 15:22:15
#> 37  2016-02-08 15:34:53
#> 38  2016-02-08 15:34:53
#> 39  2016-02-08 15:34:53
#> 40  2016-02-08 15:34:53
#> 41  2016-02-08 15:34:53
#> 42  2016-02-08 15:34:53
#> 43  2016-02-08 15:57:32
#> 44  2016-02-08 15:57:32
#> 45  2016-02-08 15:57:32
#> 46  2016-02-08 15:57:32
#> 47  2016-02-08 15:57:32
#> 48  2016-02-08 15:57:32
#> 49  2016-02-08 15:57:32
#> 50  2016-02-08 15:57:32
#> 51  2016-02-08 15:57:32
#> 52  2016-02-08 15:57:32
#> 53  2016-02-08 15:57:32
#> 54  2016-02-08 15:57:32
#> 55  2016-02-08 15:57:32
#> 56  2016-02-08 15:57:32
#> 57  2016-02-08 15:57:32
#> 58  2016-02-08 15:57:32
#> 59  2016-02-08 15:57:32
#> 60  2016-02-08 15:57:32
#> 61  2016-02-08 15:57:32
#> 62  2016-02-08 15:57:32
#> 63  2016-02-08 15:57:32
#> 64  2016-02-08 15:57:32
#> 65  2016-02-08 15:57:32
#> 66  2016-02-08 15:57:32
#> 67  2016-02-08 15:57:32
#> 68  2016-02-08 15:57:32
#> 69  2016-02-08 15:57:32
#> 70  2016-02-08 15:57:32
#> 71  2016-02-08 15:57:32
#> 72  2016-02-08 15:57:32
#> 73  2016-02-08 15:57:32
#> 74  2016-02-08 15:57:32
#> 75  2016-02-08 15:57:32
#> 76  2016-02-08 15:57:32
#> 77  2016-02-08 15:57:32
#> 78  2016-02-08 15:57:32
#> 79  2016-02-08 15:57:32
#> 80  2016-02-08 15:57:32
#> 81  2016-02-08 15:57:32
#> 82  2016-02-08 15:57:32
#> 83  2016-02-08 15:57:32
#> 84  2016-02-08 15:57:32
#> 85  2016-02-08 15:57:32
#> 86  2016-02-08 15:57:32
#> 87  2016-02-08 15:57:32
#> 88  2016-02-08 15:57:32
#> 89  2016-02-08 15:57:32
#> 90  2016-02-08 15:57:32
#> 91  2016-02-08 15:57:32
#> 92  2016-02-08 15:57:32
#> 93  2016-02-08 15:57:32
#> 94  2016-02-08 15:57:32
#> 95  2016-02-08 15:57:32
#> 96  2016-02-08 15:57:32
#> 97  2016-02-08 15:57:32
#> 98  2016-02-08 15:57:32
#> 99  2016-02-08 15:57:32
#> 100 2016-02-08 15:57:32
#> 101 2016-02-08 15:57:32
#> 102 2016-02-08 15:57:32
#> 103 2016-02-08 15:57:32
#> 104 2016-02-08 15:57:32
#> 105 2016-02-08 15:57:32
#> 106 2016-02-08 15:57:32
#> 107 2016-02-08 15:57:32
#> 108 2016-02-08 15:57:32
#> 109 2016-02-08 15:57:32
#> 110 2016-02-08 15:57:32
#> 111 2016-02-08 15:57:32
#> 112 2016-02-08 15:57:32
#> 113 2016-02-08 15:57:32
#> 114 2016-02-08 15:57:32
#> 115 2016-02-08 15:57:32
#> 116 2016-02-08 15:57:32
#> 117 2016-02-08 15:57:32
#> 118 2016-02-08 15:57:32
#> 119 2016-02-08 15:57:32
#> 120 2016-02-08 15:57:32
#> 121 2016-03-04 19:52:25
#> 122 2016-03-04 19:52:25
#> 123 2016-03-04 19:52:25
#> 124 2016-03-04 19:52:25
#> 125 2016-03-04 19:52:31
#> 126 2016-03-04 19:52:31
#> 127 2016-03-04 19:52:31
#> 128 2016-03-04 19:52:31
#> 129 2016-03-04 19:55:05
#> 130 2016-03-04 19:55:05
#> 131 2016-03-04 19:55:05
#> 132 2016-03-04 19:55:05
#> 133 2016-12-31 15:39:59
#> 134 2016-12-31 15:39:59
#> 135 2016-12-31 15:39:59
#> 136 2016-12-31 15:39:59
#> 137 2016-12-31 15:39:59
#> 138 2016-12-31 15:39:59
#> 139 2016-12-31 15:39:59
#> 140 2016-12-31 15:39:59
#> 141 2016-12-31 15:39:59
#> 142 2016-12-31 15:39:59
#> 143 2016-12-31 15:39:59
#> 144 2016-12-31 15:39:59
#> 145 2016-12-31 15:39:59
#> 146 2016-12-31 15:40:19
#> 147 2016-12-31 15:40:19
#> 148 2016-12-31 15:40:19
#> 149 2016-12-31 15:40:19
#> 150 2016-12-31 15:40:19
#> 151 2016-12-31 15:40:19
#> 152 2016-12-31 15:40:19
#> 153 2016-12-31 15:40:19
#> 154 2016-12-31 15:40:19
#> 155 2016-12-31 15:40:19
#> 156 2016-12-31 15:40:19
#> 157 2016-12-31 15:40:19
#> 158 2016-12-31 15:40:19
#> 159 2017-11-21 14:20:10
#> 160 2017-11-21 14:20:10
#> 161 2017-11-21 14:20:10
#> 162 2017-11-21 14:20:10
#> 163 2017-11-21 14:20:10
#> 164 2017-11-21 14:20:10
#> 165 2017-11-21 14:20:10
#> 166 2017-11-21 14:20:10
#> 167 2017-11-21 14:20:10
#> 168 2017-11-21 14:20:10
#> 169 2017-11-21 14:20:10
#> 170 2017-11-21 14:20:10
#> 171 2017-11-25 01:47:10
#> 172 2017-11-25 01:47:10
#> 173 2017-11-25 01:47:10
#> 174 2017-11-25 01:47:10
#> 175 2017-11-25 01:47:10
#> 176 2017-11-25 01:47:10
#> 177 2017-11-25 01:47:10
#> 178 2017-11-25 01:47:10
#> 179 2017-11-25 01:47:10
#> 180 2017-11-25 01:47:10
#> 181 2017-11-25 01:47:10
#> 182 2017-11-25 01:47:10
#> 183 2017-11-25 01:47:10
#> 184 2017-11-25 01:47:10

data(iris)
asave(iris, silent = FALSE) # iris was used in pl

#> Warning: Artifact ff575c261c949d073b2895b05d1097c3 was already archived. Another archivisation executed with success.

#> [1] "ff575c261c949d073b2895b05d1097c3"

aoptions('repoDir', NULL, unset = TRUE)

#> NULL

deleteLocalRepo(exampleRepoDir, TRUE)
rm(exampleRepoDir)